% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\name{ase}
\alias{ase}
\title{Adjacency Spectral Embedding}
\usage{
ase(A, dims)
}
\arguments{
\item{A}{matrix, an adjacency matrix that used for eigendecomposition}

\item{dims}{dimensions for which the Euclidean space is defined}
}
\value{
Adjacency Spectral Embedding
}
\description{
The eigendecomposition of an adjacency matrix provides a way to embed a graph  as points in finite dimensional Euclidean space.  This embedding allows the full arsenal of statistical and machine learning methodology  for multivariate Euclidean data to be deployed for graph inference. Adjacency Spectral Embedding performs this embedding of matrix A into dim dimensions.
}
\examples{
N <- 200
A <- matrix(data = runif(N*N), ncol = N, nrow = N)
dims <- 15
A_embed <- ase(A, dims)
}

