% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_hd.R
\name{ATM_lhd}
\alias{ATM_lhd}
\title{title ash type model for l}
\usage{
ATM_lhd(Y, Ef, Ef2, sigmae2)
}
\value{
El is the mean of l,  El2 is mean of l^2
}
\description{
description use ash type model to maxmization
}
\keyword{internal}

