% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{ATM_r1}
\alias{ATM_r1}
\title{title ash type model for l}
\usage{
ATM_r1(Y, Ef, Ef2, sigmae2, col_var = "row", nonnegative = FALSE,
  output = "mean", partype = "constant")
}
\arguments{
\item{Y}{is data matrix}

\item{Ef}{is mean for the factor}

\item{Ef2}{is second moment for the factor}

\item{sigmae2}{is variance structure for noise matrix}

\item{col_var}{is just for column variance, column is for column, row is for row}

\item{nonnegative}{is flag for whether output nonnegative value}

\item{output}{is format of output, "mean" is for mean value, "matrix" is for flash data matrix in ash}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{El}} {is a N vector for mean of loadings}
  \item{\code{El2}} {is a N vector for second moment of loadings}
 }
}
\description{
description use ash type model to maxmization
}
\keyword{internal}

