% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{Bayes_var}
\alias{Bayes_var}
\title{title Bayes variance structure estimation for kronecker productor}
\usage{
Bayes_var(sigmae2_v, sigmae2 = NULL)
}
\arguments{
\item{sigmae2_v}{residual matrix}

\item{sigmae2}{variance structure
and it is a list of two vectors in this case.}
}
\value{
sig2_out estimated variance structure
}
\description{
description prior and posterior part in objective function
}
\keyword{internal}
