% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{build_functions_df}
\alias{build_functions_df}
\title{Build dataframe of functions for input to flashcard()}
\usage{
build_functions_df(file = NULL, fs = NULL, title, desc = TRUE)
}
\arguments{
\item{file}{Character string of file name for text that includes code chunks.
Can be local file or URL.}

\item{fs}{If not using a file, character vector of functions
[do not include ()].}

\item{title}{Character string of title for flashcard deck (required)}

\item{desc}{Logical for whether to search for descriptions from
flashr_decks}
}
\value{
Dataframe suitable to include in \code{flashcard()}.
}
\description{
Build dataframe of functions for input to flashcard()
}
\examples{
build_functions_df(fs = c("apple", "apply", "+"), title = "Test")
}
