% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{change_to_mat}
\alias{change_to_mat}
\title{Check to see if a vector and if so, change to a matrix.}
\usage{
change_to_mat(x)
}
\arguments{
\item{x}{Either a vector or a matrix.}
}
\description{
Check to see if a vector and if so, change to a matrix.
}
\author{
David Gerard
}

