% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_character}
\alias{check_character}
\title{Checks if a character input is valid}
\usage{
check_character(name = NULL, value = NULL, allowed = NULL, nullok = FALSE)
}
\arguments{
\item{name}{Argument name.}

\item{value}{Argument value.}

\item{allowed}{Allowed argument values.}

\item{nullok}{Logical for whether NULL values are OK (TRUE) or not (FALSE).}
}
\description{
Checks if a character input is valid
}
\keyword{internal}
