% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_logical}
\alias{check_logical}
\title{Checks if a logical input is valid}
\usage{
check_logical(name = NULL, value = NULL)
}
\arguments{
\item{name}{Argument name.}

\item{value}{Argument value.}
}
\description{
Checks if a logical input is valid
}
\keyword{internal}
