% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_decks.R
\name{choose_deck}
\alias{choose_deck}
\title{Choose from available flashcard decks}
\usage{
choose_deck(x = NULL)
}
\arguments{
\item{x}{Name of deck to directly generate flashcards (allows choose_deck()
to act as wrapper for flashcard()).}
}
\value{
Outputs a list of available built-in flashcard decks to the console, where
the user can choose of the decks to generate flashcards.
}
\description{
Choose from available flashcard decks
}
