% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_objective.R
\name{e_loglik}
\alias{e_loglik}
\title{Get expected log likelihood under current fit}
\usage{
e_loglik(data, f)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}
}
\description{
Get expected log likelihood under current fit
}
