% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_ash}
\alias{ebnm_ash}
\title{ebnm_ash}
\usage{
ebnm_ash(x, s, ash_param, return_sampler = FALSE)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors.}

\item{ash_param}{A list of parameters to be passed into ash.}

\item{return_sampler}{Indicates whether a posterior sampler should be returned rather
than summary results.}
}
\description{
A wrapper to the ash function for flash.
}
