% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_pl}
\alias{ebnm_pl}
\title{EBNM using point-laplace prior, from ebnm package.}
\usage{
ebnm_pl(x, s, ebnm_param, return_sampler = FALSE)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors.}

\item{ebnm_param}{A list of parameters to be passed to the function
\code{ebnm_point_laplace}.}

\item{return_sampler}{This option is not yet implemented for \code{ebnm_point_laplace}.}
}
\description{
A wrapper to the function
\code{\link[ebnm]{ebnm_point_laplace}}.
}
