% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_pn}
\alias{ebnm_pn}
\title{ebnm_pn}
\usage{
ebnm_pn(x, s, ebnm_param, return_sampler = FALSE)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors.}

\item{ebnm_param}{A list of parameters to be passed to the function
\code{ebnm_point_normal}.}

\item{return_sampler}{Indicates whether a posterior sampler should be returned rather
than summary results.}
}
\description{
A wrapper to the function
  \code{\link[ebnm]{ebnm_point_normal}}.
}
