% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{eigen_plot}
\alias{eigen_plot}
\title{title eigen plot}
\usage{
eigen_plot(g, ssY)
}
\arguments{
\item{g}{greedy object which is a list}

\item{ssY}{sum square of Y}
}
\value{
return eigen value
}
\description{
description eigen plot
}
\keyword{internal}

