% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_functions}
\alias{extract_functions}
\title{Extract function calls from character vector of R code}
\usage{
extract_functions(code)
}
\arguments{
\item{code}{Object that contains R code}
}
\value{
Returns character vector of function names without parentheses (e.g.,
it returns "library" rather than "library()") included in R code.
}
\description{
Extract function calls from character vector of R code
}
\examples{
extract_functions(extract_code("https://raw.githubusercontent.com/JeffreyRStevens/flashr/refs/heads/main/README.Rmd"))
}
