% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fail_gracefully}
\alias{fail_gracefully}
\title{Gracefully fail if internet connection is not available}
\usage{
fail_gracefully(remote_file, maxtime = 10)
}
\arguments{
\item{remote_file}{Remote file to be downloaded.}

\item{maxtime}{Maximum time to check connection before timing out.}
}
\description{
CRAN policies require that "Packages which use Internet resources should fail
gracefully with an informative message if the resource is not available or
has changed (and not give a check warning nor error)." This solution is
adapted from kvasilopoulos' response at
\url{https://forum.posit.co/t/internet-resources-should-fail-gracefully/49199/11}.
\emph{This function is not exported.}
}
\keyword{internal}
