% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{find_col_blocks}
\alias{find_col_blocks}
\title{Partition a matrix into blocks of identical columns}
\usage{
find_col_blocks(X)
}
\arguments{
\item{X}{the matrix to be partitioned (note that X should not have NAs)}
}
\value{
a list, each element of which contains the indices of a single block of
identical columns
}
\description{
Partition a matrix into blocks of identical columns
}
