% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash}
\alias{flash}
\title{Main flash function}
\usage{
flash(data, Kmax, column_covariates = NULL, row_covariates = NULL,
  f_init = NULL, init_fn = NULL, var_type = c("by_column", "constant"),
  tol = 0.01, ash_param = list(), verbose = FALSE)
}
\arguments{
\item{data}{an n by p matrix or a flash data object created using \code{set_flash_data}}

\item{Kmax}{the maximum total number of factors to use (including the r1+r2 covariates)}

\item{column_covariates}{an n by r1 matrix of covariates (eg could be a column of all 1s to allow an intercept for each column)}

\item{row_covariates}{a p by r2 matrix of covariates (eg could be a vector of p 1s to allow an intercept for each row)}

\item{f_init}{a previously fitted flash object to be added to or refined}

\item{init_fn}{function used to initialize factors and loadings,}

\item{var_type}{type of variance structure to assume for residuals.}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
a fitted flash object
}
\details{
Performs Empirical Bayes factor analysis with adaptive shrinkage on both factors and loadings.
}
