% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_add_factor}
\alias{flash_add_factor}
\title{add a factor to f based on residuals}
\usage{
flash_add_factor(data, f, K = 1, init_method = c("softImpute", "svd",
  "random"))
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}
}
\details{
Computes the current residuals from data and f and adds K new factors based
on a simple initialization scheme applied to these residuals
}
