% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_data.R
\name{flash_add_factors_from_data}
\alias{flash_add_factors_from_data}
\title{Add factors to a flash object based on data}
\usage{
flash_add_factors_from_data(data, K, f_init = NULL, init_fn = "udv_si",
  backfit = TRUE, ...)
}
\arguments{
\item{K}{The number of factors to add.}

\item{f_init}{The flash object or flash fit object to which new
factors are to be added. If \code{f_init = NULL}, then a new flash
object is created.}

\item{backfit}{If \code{TRUE}, factors are refined via the backfitting
algorithm. The default behavior is to backfit all factors. Optionally,
a vector may be passed in containing the indices of factors to
backfit.}

\item{...}{Additional parameters to be passed to \code{flash_backfit}.}
}
\description{
Computes the current residuals from \code{data} and
  \code{f_init} and adds \code{K} new factors based on \code{init_fn}
  applied to these residuals. (If \code{f_init} is \code{NULL} then
  the residuals are the data.)
}
