% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_add_fixed_f}
\alias{flash_add_fixed_f}
\title{Add a set of fixed factors to a flash fit object.}
\usage{
flash_add_fixed_f(data, FF, f_init = NULL, fixf = NULL)
}
\arguments{
\item{data}{a flash data object}

\item{FF}{The factors, a p by K matrix. Missing values will be
initialized by the mean of the relevant column (but will generally
be re-estimated when refitting the model).}

\item{f_init}{A flash fit object to which factors are to be added
(if NULL then a new fit object is created).}

\item{fixf}{A p by K matrix of TRUE/FALSE values indicating which
elements of FF should be considered fixed and not changed during
updates. Default is to fix all non-missing values, so missing
values will be updated when f is updated.}
}
\value{
A flash fit object, with factors initialized from FF, and
  corresponding loadings initialized to 0..
}
