% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fixed.R
\name{flash_add_fixed_factors}
\alias{flash_add_fixed_factors}
\title{Add a set of fixed factors to a flash fit object}
\usage{
flash_add_fixed_factors(data, FF, f_init = NULL, fixf = NULL,
  init_fn = "udv_si", backfit = TRUE, ...)
}
\arguments{
\item{FF}{The factors, a p vector or p by K matrix. Missing values
will be initialized by the mean of the relevant column (but will
generally be re-estimated when refitting the model).}

\item{f_init}{The flash object or flash fit object to which new
factors are to be added. If \code{f_init = NULL}, then a new flash
object is created.}

\item{fixf}{A p by K matrix of of \code{TRUE}/\code{FALSE} values
indicating which elements of \code{FF} should be considered fixed
and not changed during updates.  The default is to fix all
non-missing values, so missing values will be updated when the
flash object is updated.}

\item{...}{Additional parameters to be passed to \code{flash_backfit}.
Note that \code{nullcheck} defaults to \code{FALSE} here.}
}
\value{
A flash object, with factors initialized from \code{FF},
  and corresponding loadings initialized to zero.
}
\description{
Add a set of fixed factors to a flash fit object
}
