% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_add_LF}
\alias{flash_add_LF}
\title{add factors or loadings to f}
\usage{
flash_add_LF(data, f, LL, FF, fixl = NULL, fixf = NULL)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{LL}{the loadings, an n by K matrix}

\item{FF}{the factors, a p by K matrix}

\item{fixl}{an n by K matrix of TRUE/FALSE values indicating which elements of LL should be considered fixed and not changed during updates.
Useful for including a mean factor for example.}

\item{fixf}{a p by K matrix of TRUE/FALSE values; same as fixl but for factors FF.}
}
\value{
a flash fit object, with additional factors initialized using LL and FF
}
\details{
The precision parameter in f is updated after adding
}
