% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_add_sparse_f}
\alias{flash_add_sparse_f}
\title{Add a set of sparse factors to a flash fit object}
\usage{
flash_add_sparse_f(data, nonnull_FF, f_init = NULL, init_fn = "udv_si",
  fix_nulls = TRUE)
}
\arguments{
\item{data}{an n by p matrix or a flash data object.}

\item{nonnull_FF}{a p by K matrix of TRUE/FALSE values. Each column corresponds to
a single sparse factor. Within each column, a value of FALSE indicates that the
respective element should be initialized to zero. Other elements are treated as
potentially non-null, and are initialized using init_fn.}

\item{f_init}{a flash fit object to which factors are to be added (if NULL then a
new fit object is created).}

\item{init_fn}{the function to use to initialize new factors (typically some kind
of svd-like function). Since only a subset of the factor elements need to be
initialized, init_fn is applied to a subsetted flash object.}

\item{fix_nulls}{indicates whether nulls should be regarded as fixed during
subsequent updates. Default is TRUE.}
}
\value{
a flash fit object, with additional sparse factors initialized using nonnull_FF
}
\examples{
sparse_ff = c(rep(5, 2), rep(0, 18))
Y = matrix(rnorm(100),nrow=5,ncol=20) + outer(rnorm(5), sparse_ff)
f = flash_add_sparse_f(Y, matrix(c(rep(TRUE, 2), rep(FALSE, 18)), ncol=1))
f = flash_backfit(Y, f, var_type="constant")
}
