% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_add_sparse_l}
\alias{flash_add_sparse_l}
\title{Add a set of sparse loadings to a flash fit object}
\usage{
flash_add_sparse_l(data, nonnull_LL, f_init = NULL, init_fn = "udv_si",
  fix_nulls = TRUE)
}
\arguments{
\item{data}{an n by p matrix or a flash data object.}

\item{nonnull_LL}{an n by K matrix of TRUE/FALSE values. Each column corresponds to
a single sparse loading. Within each column, a value of FALSE indicates that the
respective element should be initialized to zero. Other elements are treated as
potentially non-null, and are initialized using init_fn.}

\item{f_init}{a flash fit object to which factors are to be added (if NULL then a
new fit object is created).}

\item{init_fn}{the function to use to initialize new factors (typically some kind
of svd-like function). Since only a subset of the loading elements need to be
initialized, init_fn is applied to a subsetted flash object.}

\item{fix_nulls}{indicates whether nulls should be regarded as fixed during
subsequent updates. Default is TRUE.}
}
\value{
a flash fit object, with additional sparse factors initialized using nonnull_LL
}
\examples{
sparse_ll = c(rep(1, 2), rep(0, 8))
Y = matrix(rnorm(200),nrow=10,ncol=20) + outer(sparse_ll, rnorm(20))
f = flash_add_sparse_l(Y, matrix(c(rep(TRUE, 2), rep(FALSE, 8)), ncol=1))
f = flash_backfit(Y, f, var_type="constant")
}
