% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_combine}
\alias{flash_combine}
\title{combine two flash fit objects}
\usage{
flash_combine(f1, f2)
}
\arguments{
\item{f1}{first flash fit object}

\item{f2}{second flash fit object}
}
\value{
a flash fit object whose factors are concatenations of f1 and f2
The precision (tau) of the combined fit is inherited from f1
}
