% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_sampler.R
\name{flash_f_sampler}
\alias{flash_f_sampler}
\title{Generates sampler for F}
\usage{
flash_f_sampler(data, f, kset = NULL, ebnm_fn = ebnm_ash)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{kset}{the indices of factor/loadings to include when sampling LF (defaults to all)}

\item{ebnm_fn}{function used to solve the Empirical Bayes Normal Means problem}
}
\value{
A function that takes a single parameter nsamp, the number of samples of F to be
produced by the sampler. This sampler returns a list of matrices.
}
\description{
Generates function that samples F from a flash fit object. The columns of F are
sampled independently from their marginal posteriors (conditional on L being fixed at
its expectation).
}
