% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_f}
\alias{flash_get_f}
\title{flash_get_f}
\usage{
flash_get_f(f, k = NULL)
}
\arguments{
\item{f}{a flash fit object}

\item{k}{indices of factors to be returned}
}
\details{
returns the factors from a flash fit (all by default)
}
