% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_fitted_values}
\alias{flash_get_fitted_values}
\title{Get fitted values.}
\usage{
flash_get_fitted_values(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\value{
The estimated value of LF', an n by p matrix.
}
\description{
Return the estimated LF' matrix.
}
