% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_gf}
\alias{flash_get_gf}
\title{Return the list of priors on the factors}
\usage{
flash_get_gf(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\description{
Return the list of priors on the factors
}
\details{
Returns a list of length K, with the kth list element
  containing the prior for the kth factor.
}
