% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_gl}
\alias{flash_get_gl}
\title{Return the list of priors on the loadings.}
\usage{
flash_get_gl(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\description{
Return the list of priors on the loadings.
}
\details{
Returns a list of length K, with the kth list element
  containing the prior for the kth loading.
}
