% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_l}
\alias{flash_get_l}
\title{flash_get_l}
\usage{
flash_get_l(f, k = NULL)
}
\arguments{
\item{f}{a flash fit object}

\item{k}{indices of loadings to be returned}
}
\details{
returns the loadings from a flash fit (all by default)
}
