% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_ldf}
\alias{flash_get_ldf}
\title{Get LDF from a flash object}
\usage{
flash_get_ldf(f, kset = NULL, drop_zero_factors = TRUE)
}
\arguments{
\item{f}{A flash fit object.}

\item{kset}{Indices of loadings/factors to be returned.}

\item{drop_zero_factors}{If \code{TRUE}, then any factor/loadings
that are zero will be removed.}
}
\value{
A list with the following elements. These are analogous to
  the \code{u}, \code{d} and \code{v} returned by \code{svd}, but
  the columns of \code{l} and \code{f} are not orthogonal.

  \item{\code{l}}{A matrix whose columns contain the standardized
    loadings (i.e., with norm 1).}

  \item{\code{d}}{A vector of weights (analogous to the singular
    values in an SVD).}

  \item{\code{f}}{A matrix whose columns contain the standardized
    factors (i.e., with norm 1).}
}
\description{
Returns standardized loadings, factors, and weights from
a flash object.
}
