% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_lf}
\alias{flash_get_lf}
\title{Return the estimated LF' matrix.}
\usage{
flash_get_lf(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\value{
The estimated value of LF', an n by p matrix.
}
