% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_nfactors}
\alias{flash_get_nfactors}
\title{Get number of factors in a flash object}
\usage{
flash_get_nfactors(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\description{
Returns the number of factors in a flash object.
  Factors that have been zeroed out are not counted.
}
