% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_pve}
\alias{flash_get_pve}
\title{flash_get_pve}
\usage{
flash_get_pve(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\description{
Returns the factor contributions ('proportion of
variance explained') by each factor/loading combination in flash
fit f. Because the factors are not required to be orthogonal this
should be interpreted loosely: eg PVE could total more than 1.
}
