% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_udv}
\alias{flash_get_udv}
\title{flash_get_udv}
\usage{
flash_get_udv(f)
}
\details{
returns the normalized factors and weights in the same form as svd
That is, as a list, with elements u, d and v, where the columns of u and v are normalized to have length 1
}
