% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_greedy}
\alias{flash_greedy}
\title{Fit the FLASH model to data by a greedy approach}
\usage{
flash_greedy(data, Kmax, init_method = c("svd", "random"), tol = 0.01,
  ash_param = list())
}
\arguments{
\item{data}{an n by p matrix or a flash data object created using \code{set_flash_data}}

\item{Kmax}{the maximum number of factors to consider}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}
}
\value{
a fitted flash object
}
\details{
Fits the model by adding a factor and then optimizing it.
It is "greedy" in that it does not return to re-optimize previous factors.
Stops when an added factor contributes nothing, or Kmax is reached
}
\examples{
Y = matrix(rnorm(100),nrow=5,ncol=20)
data = set_flash_data(Y)
f = flash_greedy(data,10)
flash_get_sizes(f)
}
