% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_hd.R
\name{flash_hd}
\alias{flash_hd}
\title{Factor Loading Adaptive Shrinkage (heteroscedasticity version)}
\usage{
flash_hd(Y, tol = 1e-05, numtau = 500, partype = "constant",
  sigmae2 = 1)
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{tol}{which is the stop criterion for the convergence, default is 1e-5}

\item{numtau}{number of iteration, default is 500. for the backfitting case, the number of tau should be 5 or 10.}

\item{partype}{type for the noise variance, which takes values from "constant": constant variance,
"anova": variance is from anova model, "loganova": log variance is from anova model, "poisson":
variance is equal to mean, else user should provide known value for sigmae2}

\item{sigmae2}{which is a matrix for the known noise variance matrix. the default is 1 which is not useful}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a N vector for loadings}
  \item{\code{f}} {is a P vector for factors}
  \item{\code{sigmae2}} {is mean of sigma square which is estimation for the noise variance}
 }
}
\description{
flash provide rank one matrix decomposition
}
\details{
flash_hd privide rank one matrix decomposition with variational EM algorithm.
}
\examples{
N = 100
P = 200
Y = matrix(rnorm(N*P,0,1),ncol=P)
g = flash_VEM(Y)

}

