% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init}
\alias{flash_init}
\title{Initialize a flash fit object with K factors}
\usage{
flash_init(data, K = 1, method = c("svd", "random"))
}
\arguments{
\item{data}{a flash data object}

\item{K}{a number of factors to initialize}

\item{method}{indicated how to initialize: can be "svd" or "random"}
}
\value{
a flash fit object, with loadings and factors initialized
}
