% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_fromsvd}
\alias{flash_init_fromsvd}
\title{Initialize a flash fit object from a list with elements (u,d,v)}
\usage{
flash_init_fromsvd(s)
}
\arguments{
\item{s}{list with elements (u,v,d)}

\item{K}{number of factors to use}
}
\value{
a flash fit object
}
