% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_LF}
\alias{flash_init_LF}
\title{Initialize a flash fit object from the results of a factor analysis}
\usage{
flash_init_LF(LL, FF)
}
\arguments{
\item{LL}{the loadings, an n by K matrix}

\item{FF}{the factors, a p by K matrix}
}
\value{
a flash fit object, with factors initialized using L and F
}
