% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_random}
\alias{flash_init_random}
\title{Initialize a flash fit object using random N(0,1) factors}
\usage{
flash_init_random(data, K = 1)
}
\arguments{
\item{data}{a flash data object}

\item{K}{a number of factors to initialize}
}
\value{
a flash fit object, with loadings and factors initialized randomly iid N(0,1)
}
