% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_softImpute}
\alias{flash_init_softImpute}
\title{Initialize a flash fit object from data using softimpute (without penalty - lambda=0)}
\usage{
flash_init_softImpute(data, K = 1)
}
\arguments{
\item{data}{a flash data object}

\item{K}{number of factors to use}
}
\value{
a flash fit object, initialized using first K components of softimpute
}
