% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_svd}
\alias{flash_init_svd}
\title{Initialize a flash fit object from data using SVD}
\usage{
flash_init_svd(data, K = 1)
}
\arguments{
\item{data}{a flash data object}

\item{K}{number of factors to use}
}
\value{
a flash fit object, initialized using first K components of SVD
}
