% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_init_udv}
\alias{flash_init_udv}
\title{Initialize a flash fit object from a list with elements (u,d,v)}
\usage{
flash_init_udv(s, K = 1)
}
\arguments{
\item{s}{list with elements (u,v,d)}

\item{K}{the number of factors to use (factors 1:K are used)}
}
\value{
a flash fit object ready for optimization
}
