% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_sampler.R
\name{flash_l_sampler}
\alias{flash_l_sampler}
\title{Generates sampler for L}
\usage{
flash_l_sampler(data, f, kset = NULL, ebnm_fn = ebnm_pn)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{kset}{the indices of factor/loadings to include when sampling LF (defaults to all)}

\item{ebnm_fn}{function used to solve the Empirical Bayes Normal Means problem}
}
\value{
A function that takes a single parameter nsamp, the number of samples of L to be
produced by the sampler. This sampler returns a list of matrices.
}
\description{
Generates function that samples L from a flash fit object. The columns of L are
sampled independently from their marginal posteriors (conditional on F being fixed at
its expectation).
}
