% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_sampler.R
\name{flash_lf_sampler_fixedf}
\alias{flash_lf_sampler_fixedf}
\title{Generates LF sampler with F fixed at its expectation}
\usage{
flash_lf_sampler_fixedf(data, f, kset = NULL, ebnm_fn = ebnm_ash)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{kset}{the indices of factor/loadings to include when sampling LF (defaults to all)}

\item{ebnm_fn}{function used to solve the Empirical Bayes Normal Means problem}
}
\value{
A function that takes a single parameter nsamp, the number of samples of LF to be
produced by the sampler.
}
\description{
Generates function that samples LF from a flash fit object, with F fixed at its
posterior mean and the columns of L sampled independently from their marginal posteriors.
}
