% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_optimize_single_fl}
\alias{flash_optimize_single_fl}
\title{Optimize a flash factor-loading combination}
\usage{
flash_optimize_single_fl(Y, f, k, tol = 0.01)
}
\arguments{
\item{Y}{an n times p data matrix}

\item{f}{a flash object}

\item{k}{the index of the factor/loading to optmize}
}
\value{
an updated flash object
}
\details{
Iteratively updates factor and loading k of f (as well as residual precision)
to convergence of objective (used in the greedy algorithm for example)
}
