% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_factors}
\alias{flash_plot_factors}
\title{Plot the factor/loading pairs from flash results.}
\usage{
flash_plot_factors(data, f, kset = NULL, loading_label = FALSE,
  factor_label = FALSE)
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f}{A flash fit object.}

\item{kset}{The indices of the factor/loading pairs to be plotted.}

\item{loading_label}{If \code{TRUE}, then the row names of the data
matrix will be used to label the loading plots.}

\item{factor_label}{If \code{TRUE}, then the column names of the data
matrix will be used to label the factor plots.}
}
\value{
List containing:

  \item{\code{plot_f}}{A ggplot object for the factors.}

  \item{\code{plot_l}}{A ggplot object for the loadings.}
}
\description{
Plots each factor and loading as a barplot.
}
