% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_pve}
\alias{flash_plot_pve}
\title{plot a scree plot of proportion of variance explained by each factor}
\usage{
flash_plot_pve(f, main = "Scree plot of PVE for each factor")
}
\arguments{
\item{f}{the flash fit object}
}
\value{
a ggplot plot object
}
\description{
plot a scree plot of proportion of variance explained by each factor
}
