% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_r1c.R
\name{flash_r1c}
\alias{flash_r1c}
\title{Factor Loading Adaptive Shrinkage (VEM version)}
\usage{
flash_r1c(Y, tol = 1e-06, maxiter_r1 = 500)
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{tol}{which is the stop criterion for the convergence, default is 1e-5}

\item{numtau}{number of iteration, default is 500. for the backfitting case, the number of tau should be 5 or 10.}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a N vector for loadings}
  \item{\code{f}} {is a P vector for factors}
  \item{\code{sigmae2}} {is mean of sigma square which is estimation for the noise variance}
 }
}
\description{
flash provide rank one matrix decomposition
}
\details{
flash_r1 privide rank one matrix decomposition with variational EM algorithm.
}
\examples{
sim_hd = function(N, P, SF, SL, signal, a = rchisq(N,3),b = rchisq(P,1),mu = 0){
E = matrix(rep(0,N*P),nrow=N)
sig2_true = matrix(rep(0,N*P),nrow=N)
for(i in 1:N){
  for(j in 1:P){
    sig2_true[i,j] = mu + a[i] + b[j]
    E[i,j] = rnorm(1,0,sqrt(mu + a[i] + b[j]))
  }
}

K=1
lstart = rnorm(N, 0, signal)

fstart = rnorm(P, 0, signal)

index = sample(seq(1:N),(N*SL))
lstart[index] = 0
index = sample(seq(1:P),(P*SF))
fstart[index] = 0

Y = lstart \%*\% t(fstart) + E

return(list(Y = Y, L_true = lstart, F_true = fstart, Error = E,sig2_true = sig2_true))
}
N = 200
P = 500
SF = 0.5
SL = 0.5
signal = 1
data = sim_hd(N, P, SF, SL, signal, a = rep(1,N),b = rchisq(P,2),mu = 0)
sigmae2_true = data$sig2_true
Y = data$Y
E = data$Error
gc = flash_r1c(Y,tol = 1e-04,maxiter_r1 = 300)
f = gc$f
l = gc$l
sqrt(mean(((Y - l \%*\% t(f))-E)^2))/sqrt(mean((Y - E)^2))
# check the colmean
col_mean = gc$sigmae2
col_true = as.vector(sigmae2_true[1,])
plot(col_true,col_mean)
abline(a = 0,b=1,col = "red")
gc = flash_r1(Y,tol = 1e-04,maxiter_r1 = 300)
f = gc$f
l = gc$l
sqrt(mean(((Y - l \%*\% t(f))-E)^2))/sqrt(mean((Y - E)^2))
}

