% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{flash_sample}
\alias{flash_sample}
\title{sample factors and loadings from approximate posterior in f}
\usage{
flash_sample(f, n = 1)
}
\arguments{
\item{f}{a flash fit object}

\item{n}{number of samples}
}
\value{
a list of LF samples
}
\details{
As a first step just uses normal approximation to posterior with mean and variance determined by EF and EF2
}
