% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{flash_sampler}
\alias{flash_sampler}
\title{Generate posterior sampling function}
\usage{
flash_sampler(data, f, kset = NULL, fixed = c("factors", "loadings"))
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f}{A fitted flash object.}

\item{kset}{The indices of factors to be optimized (NULL indicates
all factors).}

\item{fixed}{Indicates whether to fix factors or loadings at their
posterior mean.}
}
\value{
A function that takes a single parameter \code{nsamp}, the
  number of samples of LF to be produced by the sampler. Care should
  be used when setting \code{nsamp}, because the sampler returns a
  list of matrices which are each of the same size as the data matrix.
}
\description{
Generates function that samples LF from a flash fit
  object, with either L or F fixed at its posterior mean and the
  columns of F or L sampled independently from their marginal
  posteriors.
}
