% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_utils.R
\name{flash_subset_data}
\alias{flash_subset_data}
\title{Subset a flash data object}
\usage{
flash_subset_data(data, row_subset = NULL, col_subset = NULL)
}
\arguments{
\item{row_subset}{the subset of rows to be retained}

\item{col_subset}{the subset of columns to be retained}

\item{f}{a flash fit object}
}
\value{
a subsetted flash data object
}
