% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_utils.R
\name{flash_subset_f}
\alias{flash_subset_f}
\title{Subset a flash object with respect to its factors}
\usage{
flash_subset_f(f, subset)
}
\arguments{
\item{f}{a flash fit object}

\item{subset}{the subset of factor elements to be retained}
}
\value{
a subsetted flash fit object
}
\description{
Subset a flash object with respect to its factors
}
