% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_subset_l}
\alias{flash_subset_l}
\title{Subset a flash object with respect to its loadings}
\usage{
flash_subset_l(f, subset)
}
\arguments{
\item{f}{a flash fit object}

\item{subset}{the subset of loading elements to be retained}
}
\value{
a subsetted flash fit object
}
