% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_subset_l_data}
\alias{flash_subset_l_data}
\title{Subset a flash data object with respect to the rows of Y}
\usage{
flash_subset_l_data(data, subset)
}
\arguments{
\item{subset}{the subset of rows to be retained}

\item{f}{a flash fit object}
}
\value{
a subsetted flash data object
}
