% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_utils.R
\name{flash_transpose}
\alias{flash_transpose}
\title{transpose a flash fit object}
\usage{
flash_transpose(f)
}
\arguments{
\item{f}{the flash fit object}
}
\value{
A new flash fit object, with the factors and loadings of the original flash
fit object interchanged.
}
\description{
transpose a flash fit object
}
