% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_utils.R
\name{flash_transpose_data}
\alias{flash_transpose_data}
\title{transpose a flash data object}
\usage{
flash_transpose_data(data)
}
\arguments{
\item{f}{the flash data object}
}
\value{
A new flash data object, with the matrices of the original flash data
object transposed.
}
