% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashcard.R
\name{flashcard}
\alias{flashcard}
\title{Create flashcards}
\usage{
flashcard(deck, file = NULL, termsfirst = TRUE)
}
\arguments{
\item{deck}{Name of pre-existing flashcard deck to generate}

\item{file}{Path and name of CSV file containing terms and descriptions}

\item{termsfirst}{Logical indicating whether to show terms first (TRUE) or
descriptions first (FALSE)}
}
\value{
An HTML file of terms and descriptions rendered in the RStudio viewer or
web browser.
}
\description{
The \code{flashcard()} function generates a set of flashcards with randomly
ordered pairs of terms and descriptions. The function outputs reveal.js
presentation as an HTML file. If running in RStudio, the flashcards are
output to the viewer. Otherwise, they are output to a web browser.
}
\examples{
\dontrun{
# Display terms then descriptions
flashcard(data_type)

# Display descriptions then terms
flashcard(data_type, termsfirst = FALSE)

# Display custom CSV file of terms and descriptions
flashcard(file = "data/operators.csv")
}
}
