% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashr-package.R
\docType{package}
\name{flashr-package}
\alias{flashr}
\alias{flashr-package}
\title{flashr: Create Flashcards of Terms and Definitions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for creating flashcard decks of terms and definitions. This package creates HTML slides using 'revealjs' that can be viewed in the 'RStudio' viewer or a web browser. Users can create flashcards from either existing built-in decks or create their own from CSV files or vectors of function names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/flashr}
  \item \url{https://jeffreyrstevens.github.io/flashr/}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/flashr/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
