% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{form_mean}
\alias{form_mean}
\title{Form tensor from a list of matrices of factors.}
\usage{
form_mean(factor_list)
}
\arguments{
\item{factor_list}{A list of matrices of numerics, each with the
same number of columns. The kth column of the nth element of
\code{factor_list} is the kth component of the nth mode of mean
tensor.}
}
\description{
Form tensor from a list of matrices of factors.
}
\author{
David Gerard
}

