% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{form_outer}
\alias{form_outer}
\title{Forms outer product from a list of vectors.}
\usage{
form_outer(x)
}
\arguments{
\item{x}{A list of vectors of positive numerics.}
}
\value{
An array of numerics that is the outer products of all of
    the vectors in \code{x}.
}
\description{
Forms outer product from a list of vectors.
}
\author{
David Gerard
}

