% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{Fval}
\alias{Fval}
\title{title prior and posterior part in objective function}
\usage{
Fval(mat, fit_g)
}
\arguments{
\item{mat}{matrix of flash.data in ash output which is for posterior}

\item{fit_g}{in fitted.g in ash output which is for prior}
}
\value{
PrioPost the value for the proir and posterior in objectice function
}
\description{
description prior and posterior part in objective function
}
\keyword{internal}

