% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{get_kth_column}
\alias{get_kth_column}
\title{Extract the kth column from a matrix.}
\usage{
get_kth_column(X, k)
}
\arguments{
\item{X}{A matrix.}

\item{k}{A positive integer. The column to extract.}
}
\description{
Extract the kth column from a matrix.
}
\author{
David Gerard
}

