% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_R2}
\alias{get_R2}
\title{Get the expected squared residuals from a flash data and fit object}
\usage{
get_R2(data, f)
}
\description{
Get the expected squared residuals from a flash data and fit object
}
