% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_R2k}
\alias{get_R2k}
\title{Get the expected squared residuals from a flash data and fit object excluding factor k}
\usage{
get_R2k(data, f, k)
}
\description{
Get the expected squared residuals from a flash data and fit object excluding factor k
}
