% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_Rk}
\alias{get_Rk}
\title{Get the residuals from a flash data and fit object,
excluding factor k}
\usage{
get_Rk(data, f, k)
}
\description{
Get the residuals from a flash data and fit object,
excluding factor k
}
