% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getElbows.R
\name{getElbows}
\alias{getElbows}
\title{Given a decreasingly sorted vector, return the given number of elbows}
\usage{
getElbows(d, n = 3, threshold = FALSE)
}
\arguments{
\item{d:}{the decreasingly sorted vector (e.g. a vector of standard deviations)}

\item{n:}{the number of returned elbows.}

\item{threshold:}{either FALSE or a number. If threshold is a number, then all
the elements in d that are not larger than the threshold will be ignored.}
}
\value{
a vector of length n.
}
\description{
Reference:
  Zhu, Mu and Ghodsi, Ali (2006), "Automatic dimensionality selection from
  the scree plot via the use of profile likelihood", Computational
  Statistics & Data Analysis
}
\author{
Youngser Park <youngser@jhu.edu>
}

