% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{Factor Loading Adaptive Shrinkage (K factors version using greedy algorithm)}
\usage{
greedy(Y, K, flash_para = list(), gvalue = c("lik", "eigen"))
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{K}{is the max number of factor user want. the output will provide the actual number of factor
automaticaly.}

\item{flash_para}{is the list for input parameters for flash}

\item{gvalue}{is the output style of greedy algorithm, 
"lik" provide the lowerbound of loglikelihood
"eigen" provide the sudo eigenvalue for each factor.}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a N by K matrix for loadings}
  \item{\code{f}} {is a P by K matrix for factors}
 }
}
\description{
flash provide rank K matrix decomposition with greedy algorithm
}
\details{
greedy algorithm on the residual matrix to get a rank one matrix decomposition
}
\examples{
N = 100
P = 200
Y = matrix(rnorm(N*P,0,1),ncol=P)
g = greedy(Y,10)

}

