% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{inital_Bayes_var}
\alias{inital_Bayes_var}
\title{title initial value for Bayes variance structure estimation for kronecker productor}
\usage{
inital_Bayes_var(sigmae2_v)
}
\arguments{
\item{sigmae2_v}{residual matrix}
}
\value{
sig2_out estimated variance structure
}
\description{
description initial value for Bayes variance structure estimation for kronecker productor
}
\keyword{internal}
