% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_objective.R
\name{KL_post_prior_l}
\alias{KL_post_prior_l}
\title{compute KL divergence from q (posterior approximation) to prior
for kth loading}
\usage{
KL_post_prior_l(data, f, k = 1)
}
\description{
compute KL divergence from q (posterior approximation) to prior
for kth loading
}
