% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_decks.R
\name{list_decks}
\alias{list_decks}
\title{List available built-in flashcard decks}
\usage{
list_decks(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical to prevent list information from printing to console.}
}
\value{
Outputs a list of available built-in flashcard decks to the console.
}
\description{
List available built-in flashcard decks
}
\examples{
list_decks()
}
