% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{list_prod}
\alias{list_prod}
\title{Element-wise multiplication of two lists.}
\usage{
list_prod(A, B)
}
\arguments{
\item{A}{A list.}

\item{B}{A list.}
}
\description{
Element-wise multiplication of two lists.
}
\author{
David Gerard
}

