% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{log_anova}
\alias{log_anova}
\title{title ANOVA for log residual square}
\usage{
log_anova(residualsqr)
}
\arguments{
\item{residualsqr}{residual square}
}
\value{
estimated variance structure
}
\description{
description  ANOVA for log residual square
}
\keyword{internal}

