% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_precision.R
\name{mle_precision_by_column}
\alias{mle_precision_by_column}
\title{mle for precision (separate parameter for each column)}
\usage{
mle_precision_by_column(R2)
}
\arguments{
\item{R2}{n by p matrix of squared residuals (with NAs for missing)}
}
\value{
n by p matrix of precisions (separate value for each column)
}
