% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{mle_update_modek}
\alias{mle_update_modek}
\title{In MLE algorithm for mean zero diagonal Kronecker structured
variance model, update mode k.}
\usage{
mle_update_modek(R, esig_list, k)
}
\arguments{
\item{R}{An array of numerics.}

\item{esig_list}{A list of current values of the variances.}

\item{k}{The current mode to update}
}
\description{
In MLE algorithm for mean zero diagonal Kronecker structured
variance model, update mode k.
}
\author{
David Gerard
}

