% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_objective.R
\name{NM_posterior_e_loglik}
\alias{NM_posterior_e_loglik}
\title{expected log likelihood for normal means model}
\usage{
NM_posterior_e_loglik(x, s, Et, Et2)
}
\arguments{
\item{x}{observations in normal means}

\item{s}{standard errors of x}

\item{Em}{posterior mean of theta}

\item{Em2}{posterior second moment of theta}
}
\description{
expected log likelihood for normal means model
}
\details{
The likelihood is for x | theta ~ N(theta, s^2);
The expectation is taken over the posterior on theta
}
