% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{noisy_var_column}
\alias{noisy_var_column}
\title{title noisy variance structure estimation with noisy variance on each column}
\usage{
noisy_var_column(sigmae2_v, sigmae2_true)
}
\arguments{
\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{variance structure
and it is a list of two vectors in this case.}
}
\value{
sig2_out estimated variance structure
}
\description{
description prior and posterior part in objective function
}
\keyword{internal}
