% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{obj}
\alias{obj}
\title{title objective function in VEM}
\usage{
obj(N, P, sigmae2_v, sigmae2_true, par_f, par_l, objtype = "margin_lik")
}
\arguments{
\item{N}{dimension of residual matrix}

\item{P}{dimension of residual matrix}

\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{true variance structure (we use the estimated one to replace that if the truth is unknown)}

\item{par_f}{ash output for f}

\item{par_l}{ash output for l}

\item{objtype}{objective function type, 
"margin_lik" for conditional likelihood,
"lowerbound_lik" for full objective function}
}
\value{
obj_val value of the objectice function
}
\description{
description  objective function
}
\keyword{internal}

