% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{perform_nullcheck}
\alias{perform_nullcheck}
\title{Check whether zeroing out a factor improves the objective}
\usage{
perform_nullcheck(data, f, k, var_type, verbose)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash object}

\item{k}{the index of the factor/loading to optimize}

\item{var_type}{type of variance structure to assume for residuals.}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
a flash object
}
\details{
If zeroing out factor k improves the objective then returns
flash object with factor k set to 0 (and precision updated);
otherwise returns original flash object
}
