% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.flash}
\alias{plot.flash}
\title{Plot flash object}
\usage{
\method{plot}{flash}(x, plot_scree = TRUE, plot_factors = FALSE,
  factor_kset = 1:x$nfactors, factor_colors = NULL,
  factor_legend_size = 5, plot_loadings = FALSE,
  loading_kset = 1:x$nfactors, loading_colors = NULL,
  loading_legend_size = 5, plot_grid_nrow = 2, plot_grid_ncol = 2,
  ask = (plot_factors || plot_loadings) && dev.interactive(), ...)
}
\arguments{
\item{x}{Flash object to plot.}

\item{plot_scree}{Whether to include a scree plot of the proportion
of variance explained per factor/loading pair.}

\item{plot_factors}{Whether to plot the factors indexed by
\code{factor_kset}.}

\item{factor_kset}{If \code{plot_factors} is \code{TRUE}, then
\code{factor_kset} specifies which factors will be plotted.
Defaults to all factors.}

\item{factor_colors}{If \code{plot_factors} is \code{TRUE}, then
\code{factor_colors} specifies the colors to be used for the factor
variables. If colors are used, then a legend will be shown
alongside the factor plots.}

\item{factor_legend_size}{If \code{factor_colors} is not \code{NULL},
then \code{factor_legend_size} specifies the size of the legend
show alongside the factor plots.}

\item{plot_loadings}{Whether to plot the loadings indexed by
\code{loading_kset}.}

\item{loading_kset}{If \code{plot_loadings} is \code{TRUE}, then
\code{loading_kset} specifies which loadings will be plotted.
Defaults to all loadings.}

\item{loading_colors}{If \code{plot_loadings} is \code{TRUE}, then
\code{loading_colors} specifies the colors to be used for the
loading variables. If colors are used, then a legend will be shown
alongside the loading plots.}

\item{loading_legend_size}{If \code{loading_colors} is not \code{NULL},
then \code{loading_legend_size} specifies the size of the legend
show alongside the loading plots.}

\item{plot_grid_nrow}{The number of rows to use when displaying
multiple factor/loading plots on a single screen.}

\item{plot_grid_ncol}{The number of columns to use when displaying
multiple factor/loading plots on a single screen.}

\item{ask}{Should the user be prompted before displaying each
successive plot?}

\item{...}{Additional arguments (not used by this method).}
}
\description{
\code{plot} method for class \code{'flash'}.
}
