% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_factors}
\alias{plot_factors}
\title{plot the factors from flash results}
\usage{
plot_factors(data, f, k = 1, loading_label = FALSE, factor_label = FALSE)
}
\arguments{
\item{data}{the flash data object}

\item{f}{is a flash fit object}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{plot_f}} {is a ggplot object for the factors}
  \item{\code{plot_l}} {is a ggplot object for the loadings}
 }
}
\details{
plots each factor and loading as a barplot.
}
