% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_one_factor}
\alias{plot_one_factor}
\title{plot_f}
\usage{
plot_one_factor(f, pve, k, f_labels = NA, y_lab = "factor values")
}
\arguments{
\item{f}{factor to plot}

\item{pve}{pve for this factor}

\item{k}{the order of the factor}

\item{f_labels}{the lables for the factor}

\item{y_lab}{is the name of the Y axis}

\item{P}{the lenght of this factor}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{plot_f}} {is a ggplot object for the factors}
 }
}
\description{
plot the factor
}
\details{
plot_f is to plot the factor as barplot.
}
