% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1_opt.R
\name{r1_opt}
\alias{r1_opt}
\title{r1_opt}
\usage{
r1_opt(R, R2, l_init, f_init, l2_init = NULL, f2_init = NULL,
  l_subset = 1:length(l_init), f_subset = 1:length(f_init),
  ebnm_fn = ebnm_ash, ebnm_param = flash_default_ebnm_param(ebnm_fn),
  var_type = c("by_column", "constant", "by_row", "kroneker"), tol = 0.001,
  calc_F = TRUE, missing = NULL, verbose = FALSE, maxiter = 5000)
}
\arguments{
\item{R}{an n times p matrix of data (expected residuals)}

\item{R2}{an n times p matrix of expected squared residuals}

\item{l_init}{the initial value of loading used for iterative scheme (n vector)}

\item{f_init}{the initial value of the factor for iterative scheme (p vector)}

\item{l2_init}{initial value of l2 (optional)}

\item{f2_init}{initial value of f2 (optional)}

\item{l_subset}{vector of indices of l to update (default is all)}

\item{f_subset}{vector of indices of f to update (default is all)}

\item{ebnm_fn}{function to solve the Empirical Bayes normal means problem}

\item{ebnm_param}{parameters to be passed to ebnm_fn when optimizing}

\item{var_type}{the type of variance structure to assume}

\item{tol}{a tolerance on changes in l and f to diagnose convergence}

\item{calc_F}{whether to compute the objective function (useful for testing purposes)}

\item{missing}{an n times matrix of TRUE/FALSE indicating which elements of R and R2 should be considered missing (note neither R nor R2 must have missing values; eg set them to 0)}

\item{verbose}{if true then trace of objective function is printed}

\item{maxiter}{an upper bound on the number of iterations before terminating}
}
\value{
an updated flash object
}
\description{
Optimize a single loading and factor ("rank 1" model)
}
\details{
This function iteratively optimizes the loading, factor and residual precision, from residuals and their expected squared values
Currently the tolerance is on the changes in l and f (not on the objective function)
}
