% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computerank.R
\name{rank_matrices}
\alias{rank_matrices}
\title{Matrix Rank Computer}
\usage{
rank_matrices(graphs, normalize = FALSE)
}
\arguments{
\item{graphs, }{the matrix that will be ranked}
}
\value{
rg, the rank of the matrix
}
\description{
The vector space dimension of a matrix spanned by its columns
}
\examples{
d <- dim(graphs)
rg <- array(rep(NaN, d[1]*d[2]*d[3]), d)
for (i in 1:d[3]) {
rg[,,i] <- array(rank(graphs[,,i], ties.method="average"), c(d[1], d[2]))
if (normalize) {
rg[,,i] <- ( rg[,,i] - min(rg[,,i]) ) / (max(rg[,,i]) - min(rg[,,i]))
}
}

}

