% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{replace_factors}
\alias{replace_factors}
\title{Replace the \code{k}th columns in the matrices in \code{factor_list} with the
vectors in \code{new_factors}.}
\usage{
replace_factors(factor_list, new_factors, k)
}
\arguments{
\item{factor_list}{A list of matrices of numerics, each with the
same number of columns. The kth column of the nth element of
\code{factor_list} is the kth component of the nth mode of mean
tensor.}

\item{new_factors}{A list of vectors. The new factors to place in
the \code{k}th column of \code{factor_list}.}

\item{k}{A positive integer. The column of the matrices in
\code{factor_list} to replace.}
}
\description{
Replace the \code{k}th columns in the matrices in \code{factor_list} with the
vectors in \code{new_factors}.
}
\author{
David Gerard
}

