% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{rescale_factors}
\alias{rescale_factors}
\title{Rescale factors so some don't get too big.}
\usage{
rescale_factors(x)
}
\arguments{
\item{x}{A list of vectors.}
}
\description{
Rescale factors so some don't get too big.
}
\author{
David Gerard
}

