% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{rescale_sigmae2}
\alias{rescale_sigmae2}
\title{title rescale the lambda_l and lambda_f for identifiablity}
\usage{
rescale_sigmae2(sig2_l, sig2_f)
}
\arguments{
\item{sig2_l}{variance of the kronecker product}

\item{sig2_l}{variance of the kronecker product}
}
\value{
a list of sig2_l and sig2_f for the rescaled variance of the kronecker product
}
\description{
description rescale the lambda_l and lambda_f for identifiablity
}
\keyword{internal}
