% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_sampler.R
\name{sampler}
\alias{sampler}
\title{Generates sampler for a single factor/loading}
\usage{
sampler(is_fixed, sample_fxn, fixed_vals)
}
\arguments{
\item{is_fixed}{a vector of Booleans that indicates which elements are fixed}

\item{sample_fxn}{A function that returns samples for non-fixed elements. The function
should return a matrix whose rows correspond to individual samples.}

\item{fixed_vals}{the values of the fixed elements}
}
\value{
A function that generates samples for the factor/loading. This function returns
a matrix whose rows correspond to individual samples.
}
