% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{sigma_est}
\alias{sigma_est}
\title{title module for estiamtion of the variance structure}
\usage{
sigma_est(sigmae2_v, sigmae2_true, partype = "constant")
}
\arguments{
\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{true variance structure (we use the estimated one to replace that if the truth is unknown)}

\item{partype}{parameter type for the variance, 
"constant" for constant variance, 
"var_col" for nonconstant variance for column, 
"known" for the kown variance,
"Bayes_var" for Bayes version of the nonconstant variance for row and column
"loganova" is anova estiamtion for the log residual square}
}
\value{
sigmae2 estimated variance structure
}
\description{
description estiamtion of the variance structure
}
\keyword{internal}

