% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{sigmae2_v_est}
\alias{sigmae2_v_est}
\title{title empirical sample variance matrix estimation}
\usage{
sigmae2_v_est(Y, El, Ef, El2, Ef2, fl_list = list())
}
\arguments{
\item{Y}{which is the residual}

\item{El}{mean estimation for the current loading}

\item{Ef}{mean estimation for the current factor}

\item{El2}{second momnet estimation for the current factor}

\item{Ef2}{second moment estimation for the current loading}

\item{fl_list}{this is a list for the El Ef El2 and Ef2 from other factors}
}
\value{
sigmae2_v  empirical sample variance matrix
}
\description{
description this function is to calculate the empirical sample variance matrix sigmae2_v
}
\keyword{internal}

