% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.flash}
\alias{summary.flash}
\title{Summarize flash object}
\usage{
\method{summary}{flash}(object, data, ...)
}
\arguments{
\item{object}{A flash fit object.}

\item{data}{A matrix or flash data object created using
\code{flash_set_data}.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
A list with the following elements.

  \item{nfactors}{The number of factors in the flash object. See
    \code{\link{flash_get_nfactors}}.}

  \item{pve}{The proportion of variance explained by each factor. See
    \code{\link{flash_get_pve}}.}

  \item{fitted.values}{The estimated LF' matrix. See
    \code{\link{flash_get_fitted_values}}.}

  \item{ldf}{Standardized loadings, factors, and weights. See
    \code{\link{flash_get_ldf}}.}

  \item{objective}{The value of the objective function attained by the
    fit. See \code{\link{flash_get_objective}}.}
}
\description{
\code{summary} method for class \code{'flash'}.
}
