% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_est.R
\name{tau_est}
\alias{tau_est}
\title{title module for estiamtion of the variance structure}
\usage{
tau_est(f, data)
}
\arguments{
\item{partype}{parameter type for the variance,
"constant" for constant variance,
"var_col" for nonconstant variance for column,
"known" for the kown variance,
"Bayes_var" for Bayes version of the nonconstant variance for row and column
"noisy" is noisy version s_ij + sigmae2
"noisy_col" noisy version for column s_ij + sigmae2_j}

\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{true variance structure}

\item{sigmae2_pre}{the previouse sigmae2 which is only used in Bayes_var}
}
\value{
sigmae2 estimated variance structure
}
\description{
description estiamtion of the variance structure
}
\keyword{internal}
