% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{tbackfitting}
\alias{tbackfitting}
\title{Perform backfitting starting at output of \code{tgreedy}.}
\usage{
tbackfitting(Y, factor_list, sig_vec, maxiter_bf = 100, tol_bf = 10^-6,
  maxiter_vem = 100, mixcompdist = "normal", alpha = 0, beta = 0)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{factor_list}{A list of matrices with the same number of
columns. These are the starting values for the backfitting
algorithm. The intended starting values are can be obtained
from \code{\link{tgreedy}}.}

\item{sig_vec}{A vector of positive numerics. The estimates of
variances that were returned by \code{\link{tgreedy}}.}

\item{maxiter_bf}{A positive integer. The maximum number of
backfitting steps to perform.}

\item{tol_bf}{A positive numeric. The stopping criterion for the
backfitting algorithm.}

\item{maxiter_vem}{A positive integer. The maximum number of steps
in each VEM algorithm to perform at each iteration of the
backfitting algorithm.}

\item{mixcompdist}{What should the mixing distribution be? See
options from the `ashr` package.}

\item{alpha}{The prior shape parameter for the variance.}

\item{beta}{The prior rate parameter for the variance.}
}
\description{
Perform backfitting starting at output of \code{tgreedy}.
}
\author{
David Gerard
}

