% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{tflash}
\alias{tflash}
\title{Tensor Factor Loading Adaptive SHrinkage (T-FLASH).}
\usage{
tflash(Y, var_type = c("homoscedastic", "kronecker"), tol = 10^-5,
  itermax = 100, alpha = 0, beta = 0, mixcompdist = "normal",
  sig_start_itermax = 10, nullweight = 10, print_update = FALSE,
  start = c("first_sv", "random"))
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{var_type}{A string. What variance model should we assume?
Options are homoscedastic noise (\code{"homoscedastic"}) or
Kronecker structured variance (\code{kronecker}).}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{sig_start_itermax}{A positive integer. The number of
iterations to run in initializing the precision before starting
the VEM. Defaults to 10.}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of zero.}

\item{print_update}{A logical. Should we print notifications on how
far along the optimization is?}

\item{start}{How should we choose the starting values? Either using
the first singular vector along each mode (\code{"first_sv"})
or randomly (\code{"random"}).}
}
\description{
Fits a rank-1 tensor mean model with a homoscedastic error term. It
does this via a variational Bayesian approach with a unimodal prior
on the components of the mean tensor.
}
\author{
David Gerard
}

