% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{tflash_homo}
\alias{tflash_homo}
\title{Homoscedastic variance implementation of T-FLASH.}
\usage{
tflash_homo(Y, tol = 10^-5, itermax = 100, alpha = 0, beta = 0,
  mixcompdist = "normal", sig_start_itermax = 10, nullweight = 1)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{sig_start_itermax}{A positive integer. The number of
iterations to run in initializing the precision before starting
the VEM. Defaults to 10.}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of zero.}
}
\description{
Homoscedastic variance implementation of T-FLASH.
}
\author{
David Gerard
}

