% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tflash_kron}
\alias{tflash_kron}
\title{Tensor FLASH assuming a diagonal Kronecker structured covariance model.}
\usage{
tflash_kron(Y, tol = 10^-5, itermax = 100, alpha = 0, beta = 0,
  mixcompdist = "normal", nullweight = 10, print_update = FALSE,
  start = c("first_sv", "random"), known_modes = NULL,
  known_factors = NULL, homo_modes = NULL)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of zero.}

\item{print_update}{A logical. Should we print notifications on how
far along the optimization is?}

\item{start}{How should we choose the starting values? Either using
the first singular vector along each mode (\code{"first_sv"})
or randomly (\code{"random"}).}

\item{known_modes}{A vector of integers. The modes that are
known. Should be the same length as \code{known_factors}.}

\item{known_factors}{A list of known factors for the modes
indicated in \code{known_modes}. Defaults to \code{NULL}, where
all factors are assumed to be unknown.}

\item{homo_modes}{A vector of integers. If \code{var_type =
"kronecker"} then \code{homo_modes} indicates which modes are
assumed to be homoscedastic.}
}
\description{
Tensor FLASH assuming a diagonal Kronecker structured covariance model.
}
\author{
David Gerard
}

