% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{tgreedy}
\alias{tgreedy}
\title{Iteratively do T-FLASH on the residuals.}
\usage{
tgreedy(Y, k = max(dim(Y)), tol = 10^-5, itermax = 100, alpha = 0,
  beta = 0, mixcompdist = "normal")
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{k}{The maximum cp-rank of the mean tensor.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}
}
\description{
This function will estimate a rank-1 tensor using independent
unimodal priors on the components via a variational EM
algorithm. It will subtract off the posterior mean from the data
array, then repeat the estimation procedure on the residuals. It
continues to do this until the maximum cp-rank \code{k} is reached
or when all of the components are estimated to be 0.
}
\author{
David Gerard
}
\seealso{
\code{\link{tflash}} for fitting the rank-1 mean tensor
    model.
}

