% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{tinit_components}
\alias{tinit_components}
\title{Obtain initial estimates of each mode's components.}
\usage{
tinit_components(Y, which_na = NULL)
}
\arguments{
\item{Y}{An array of numerics.}

\item{which_na}{Either NULL (when complete data) or an array of
logicals the same dimension as \code{Y}, indicating if the
observation is missing (\code{TRUE}) or observed
(\code{FALSE}).}
}
\value{
\code{ex_list} A list of vectors of the starting expected
    values of each component.


    \code{ex2_vec} A vector of starting values of \eqn{E[x'x]}
}
\description{
The inital values of each component are taken to be proportional to
the first singular vector of each mode-k matricization of the data
array.
}
\author{
David Gerard
}

