% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tupdate_kron_modek}
\alias{tupdate_kron_modek}
\title{Update the variational density for the kth mode when assuming a
diagonal Kronecker structured covariance matrix.}
\usage{
tupdate_kron_modek(Y, ex_list, ex2_list, esig_list, k, mixcompdist = "normal")
}
\arguments{
\item{Y}{An array of numerics. The data array.}

\item{ex_list}{A list of vectors of numerics. The expected values
of the components of each mode.}

\item{ex2_list}{A list of of vectors of positive numerics. The
variational second moments of the components.}

\item{esig_list}{A list of vectors of positive numerics. The
expected variances.}

\item{k}{A positive integer. The current mode to update.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}
}
\description{
Update the variational density for the kth mode when assuming a
diagonal Kronecker structured covariance matrix.
}
\author{
David Gerard
}

