% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tupdate_kron_sig}
\alias{tupdate_kron_sig}
\title{Update the kth mode diagional Kronecker structured covariance
matrix.}
\usage{
tupdate_kron_sig(Y, ex_list, ex2_list, esig_list, k, beta = 0,
  which_na = NULL)
}
\arguments{
\item{Y}{An array of numerics. The data array.}

\item{ex_list}{A list of vectors of numerics. The expected values
of the components of each mode.}

\item{ex2_list}{A list of of vectors of positive numerics. The
variational second moments of the components.}

\item{esig_list}{A list of vectors of positive numerics. The
expected variances.}

\item{k}{A positive integer. The current mode to update.}

\item{beta}{A non-negative vector of numerics. Prior rate
parameters for the mode-k variances.}

\item{which_na}{Either NULL (when complete data) or an array of
logicals the same dimension as \code{Y}, indicating if the
observation is missing (\code{TRUE}) or observed
(\code{FALSE}).}
}
\value{
\code{post_rate} The variational posterior rate parameter.
}
\description{
Update the kth mode diagional Kronecker structured covariance
matrix.
}

