% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{tupdate_modek}
\alias{tupdate_modek}
\title{Update the mode k variational density.}
\usage{
tupdate_modek(Y, ex_list, ex2_vec, esig, k, mixcompdist = "normal",
  which_na = NULL, nullweight = 1)
}
\arguments{
\item{Y}{An array of numerics. The data array.}

\item{ex_list}{A list of vectors of numerics. The expected values
of the components of each mode.}

\item{ex2_vec}{A vector of positive numerics. The expected value of
\eqn{x'x} for each mode.}

\item{esig}{A positive numeric. The expected value of the
precision.}

\item{k}{A positive integer. The current mode to update.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{which_na}{Either NULL (when complete data) or an array of
logicals the same dimension as \code{Y}, indicating if the
observation is missing (\code{TRUE}) or observed
(\code{FALSE}).}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of a component being zero.}
}
\value{
\code{ex_list} A list of vectors of the starting expected
    values of each component.

    \code{ex2_vec} A vector of starting values of \eqn{E[x'x]}
}
\description{
Update the mode k variational density.
}
\author{
David Gerard
}

