% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_random}
\alias{udv_random}
\title{udv_random}
\usage{
udv_random(Y, K = 1)
}
\arguments{
\item{Y}{An n by p matrix.}

\item{K}{Number of factors.}
}
\value{
A list with components (u,d,v), with elements of u and v
  i.i.d. N(0,1).
}
\description{
Provides a random initialization of factors.
}
