% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_si}
\alias{udv_si}
\title{udv_si}
\usage{
udv_si(Y, K = 1)
}
\arguments{
\item{Y}{An n by p matrix.}

\item{K}{Number of factors to use.}
}
\value{
A list with components (u,d,v).
}
\description{
Provides a simple wrapper to \code{softImpute} to
  provide a rank 1 initialization. Uses \code{type = "als"} option.
}
