% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_si_svd}
\alias{udv_si_svd}
\title{udv_si}
\usage{
udv_si_svd(Y, K = 1)
}
\arguments{
\item{Y}{an n by p matrix}

\item{K}{number of factors to use}
}
\value{
A list with components (u,d,v)
}
\description{
provides a simple wrapper to \code{softImpute} to
  provide a rank 1 initialization. Uses \code{type = "svd"} option.
}
