% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_svd}
\alias{udv_svd}
\title{udv_svd}
\usage{
udv_svd(Y, K = 1)
}
\arguments{
\item{Y}{An n by p matrix.}

\item{K}{Number of factors to use.}
}
\value{
A list with components (u,d,v).
}
\description{
Provides a simple wrapper to svd.
}
