% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\docType{data}
\name{vectors}
\alias{vectors}
\title{Vectors deck}
\format{
A data frame with 4 columns.
\describe{
\item{term}{reference term or function}
\item{description}{description or definition of term}
\item{package}{package including function/argument}
\item{title}{title of deck}
}
}
\usage{
vectors
}
\description{
This flashcard deck includes terms associated with vectors.
}
\concept{decks}
\keyword{datasets}
