% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfitting.R
\name{backfitting}
\alias{backfitting}
\title{backfitting for correction K factors model}
\usage{
backfitting(Y, Lest, Fest, tautol = 100, numtau = 500)
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{Lest}{is estimation for l to correct}

\item{Fest}{is estimation for f to correct}

\item{tautol}{this the stop criterion for convergence, number for iteration}

\item{numtau}{number for iteration in flash rank one.}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a K by N matrix for loadings}
  \item{\code{f}} {is a K by P matrix for factors}
 }
}
\description{
correction the factor and loading matrix estimation using backfitting algorithm
}
\details{
greedy algorithm on the residual matrix to get a rank one matrix decomposition
}
\examples{
N = 100
P = 200
Y = matrix(rnorm(N*P,0,1),ncol=P)
g = flash_v_K(Y,10)

}

