% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfitting.R
\name{backfitting}
\alias{backfitting}
\title{backfitting to correct K-factor model}
\usage{
backfitting(Y, Lest, Fest, maxiter_bf = 100, maxiter_r1 = 500,
  r1_type = "constant")
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{Lest}{is estimate for l to correct}

\item{Fest}{is estimate for f to correct}

\item{maxiter_bf}{maximum number of iterations}

\item{maxiter_r1}{maximum number of iterations in flash rank one.}

\item{r1_type}{is the flag to choose the constant variance for columns or not, defulat is constant}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a N by K matrix for loadings}
  \item{\code{f}} {is a P by K matrix for factors}
  \item{\code{sigmae2}} {is variance for the error, scalar for constant case and vector for nonconstant case}
 }
}
\description{
Correct the factor and loading matrix estimates using backfitting algorithm
}
\details{
Repeatedly applies rank 1 algorithm to Y-L[,-i]F[,-i]'
}
\examples{
N = 100
P = 200
Y = matrix(rnorm(N*P,0,1),ncol=P)
g = greedy(Y,K = 10)
gb = backfitting(Y,g$l,g$f,maxiter_bf=100,maxiter_r1 = 5)

}

