% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfitting.R
\name{backfitting}
\alias{backfitting}
\title{backfitting to correct K-factor model}
\usage{
backfitting(Y, initial_list, maxiter_bf = 100, flash_para = list(),
  gvalue = c("lik", "eigen"), parallel = FALSE)
}
\arguments{
\item{Y}{is the data matrix (N by P)}

\item{initial_list}{is the list from initial_list algorithm}

\item{maxiter_bf}{maximum number of iterations}

\item{gvalue}{is for output of the backfit,
"eigen" means just provide the sum of square
"lik" mean provide the lowerbound}

\item{Fest}{is estimate for f to correct}

\item{flash_parais}{the list for flash parameters setting up}

\item{initial_list}{initial list for the starting value which contains
l,f,l2,f2
priorpost_vec
clik_vec}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{l}} {is a N by K matrix for loadings}
  \item{\code{f}} {is a P by K matrix for factors}
  \item{\code{sigmae2}} {is variance for the error, scalar for constant case and vector for nonconstant case}
 }
}
\description{
Correct the factor and loading matrix estimates using backfitting algorithm
}
\details{
Repeatedly applies rank 1 algorithm to Y-L[,-i]F[,-i]'
}
\examples{
N = 100
P = 200
Y = matrix(rnorm(N*P,0,1),ncol=P)
g = initial_list(Y,K = 10)
gb = backfitting(Y,g$l,g$f,maxiter_bf=100,maxiter_r1 = 5)

}

