% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{Bayes_var}
\alias{Bayes_var}
\title{title Bayes variance structure estimation for kronecker productor}
\usage{
Bayes_var(sigmae2_v, sigmae2_true)
}
\arguments{
\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{true variance structure (we use the estimated one to replace that if the truth is unknown)}
}
\value{
sig2_out estimated variance structure
}
\description{
description prior and posterior part in objective function
}
\keyword{internal}

