% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{C_likelihood}
\alias{C_likelihood}
\title{title conditional likelihood}
\usage{
C_likelihood(N, P, sigmae2_v, sigmae2_true)
}
\arguments{
\item{N}{dimension of residual matrix}

\item{P}{dimension of residual matrix}

\item{sigmae2_v}{residual matrix}

\item{sigmae2_true}{true variance structure (we use the estimated one to replace that if the truth is unknown)}
}
\value{
c_lik for the onditional likelihood in objectice function
}
\description{
description conditional likelihood in objective function
}
\keyword{internal}

