% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{Data types deck}
\format{
A data frame with 4 columns.
\describe{
\item{term}{reference term or function}
\item{description}{description or definition of term}
\item{url}{URL for function documentation}
\item{package}{package including function/argument}
\item{title}{title of deck}
}
}
\usage{
data_types
}
\description{
This flashcard deck includes terms associated with data types and structures.
}
\concept{decks}
\keyword{datasets}
