% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{diag_mle}
\alias{diag_mle}
\title{MLE of mean zero Kronecker variance model.}
\usage{
diag_mle(R, itermax = 100, tol = 10^-3)
}
\arguments{
\item{R}{An array of numerics.}
}
\value{
\code{esig_list} A list of vectors of numerics. The MLEs of
    the precisions, not the variances.
}
\description{
MLE of mean zero Kronecker variance model.
}
\author{
David Gerard
}

