% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_ash}
\alias{ebnm_ash}
\title{provides functions to solve the Empirical Bayes Normal Means problem
function must take the arguments x,s and output a list with elements
postmean, postmean2, fitted_g, penloglik}
\usage{
ebnm_ash(x, s, ash_param)
}
\description{
provides functions to solve the Empirical Bayes Normal Means problem
function must take the arguments x,s and output a list with elements
postmean, postmean2, fitted_g, penloglik
}
