% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_ash}
\alias{ebnm_ash}
\title{A wrapper to the ash function for flash.}
\usage{
ebnm_ash(x, s, ash_param)
}
\arguments{
\item{x}{Describe parameter here.}

\item{s}{Describe parameter here.}

\item{ash_param}{Describe parameter here.}
}
\description{
A wrapper to the ash function for flash.
}
