% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_pl}
\alias{ebnm_pl}
\title{ebnm_pl}
\usage{
ebnm_pl(x, s, ebnm_param)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors.}

\item{ebnm_param}{A list of parameters to be passed to the function
ebnm_point_laplace.}
}
\description{
A wrapper to the function \code{\link[ebnm]{point_laplace}}.
}
\details{
EBNM using point-laplace prior, from ebnm package.
}
