% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_functions.R
\name{ebnm_pl}
\alias{ebnm_pl}
\title{ebnm_pl}
\usage{
ebnm_pl(x, s, ebnm_param)
}
\arguments{
\item{x}{a vector of observations}

\item{s}{a vector of standard errors}

\item{ebnm_param}{a list of parameters to be passed to the function ebnm_point_laplace}
}
\details{
a wrapper to the function ebnm::point_laplace
}
