% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash}
\alias{flash}
\title{FLASH}
\usage{
flash(Y, tol = 1e-05, maxiter_r1 = 500, partype = c("constant", "known",
  "Bayes_var", "var_col"), sigmae2_true = NA, factor_value = NA,
  fix_factor = FALSE, nonnegative = FALSE, objtype = c("margin_lik",
  "lowerbound_lik"), ash_para = list())
}
\arguments{
\item{Y}{the data matrix}

\item{tol}{is for the tolerence for convergence in iterations and ash}

\item{partype}{parameter type for the variance,
"constant" for constant variance,
"var_col" for nonconstant variance for column,
"known" for the kown variance,
"Bayes_var" for Bayes version of the nonconstant variance for row and column
"loganova" is anova estiamtion for the log residual square}

\item{sigmae2_true}{true value for the variance structure}

\item{factor_value}{is the factor value if the factor is fixed}

\item{fix_factor}{whether the factor is fixed or not
TRUE for fix_factor
FALSE for non-constraint}

\item{nonnegative}{if the facotor and loading are nonnegative or not.
TRUE for nonnegative
FALSE for no constraint}

\item{objtype}{objective function type,
"margin_lik" for conditional likelihood,
"lowerbound_lik" for full objective function}

\item{maciter_r1}{is maximum of the iteration times for rank one case}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{El}} {is a N vector for mean of loadings}
  \item{\code{Ef}} {is a N vector for mean of factors}
  \item{\code{sigmae2}}{is a N by P matrix for estimated value for the variance structure}
 }
}
\description{
factor loading adaptive shrinkage rank one version
}
\details{
flash privide rank one matrix decomposition with variational EM algorithm.
}

