% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fixed.R
\name{flash_add_fixed_loadings}
\alias{flash_add_fixed_loadings}
\title{Add a set of fixed loadings to a flash fit object}
\usage{
flash_add_fixed_loadings(data, LL, f_init = NULL, fixl = NULL, var_type,
  ebnm_fn, ebnm_param, stopping_rule, tol, verbose_output, nullcheck, maxiter)
}
\arguments{
\item{LL}{The loadings, an n by K matrix. Missing values will be
initialized by the mean of the relevant column (but will generally
be re-estimated when refitting the model).}

\item{f_init}{The flash object or flash fit object to which new
factors are to be added. If \code{f_init = NULL}, then a new flash
object is created.}

\item{fixl}{An n by K matrix of \code{TRUE}/\code{FALSE} values
indicating which elements of \code{LL} should be considered fixed
and not changed during updates.  The default is to fix all
non-missing values, so missing values will be updated when the
flash object is updated.}

\item{var_type}{The type of variance structure to assume for
residuals. Options include:
\describe{
  \item{\code{"by_column"}}{Residuals in any given column are
    assumed to have the same variance.}
  \item{\code{"by_row"}}{Residuals in any given row have the
    same variance.}
  \item{\code{"constant"}}{All residuals are assumed to have the
    same variance.}
  \item{\code{"zero"}}{The variance of the residuals is fixed. To
    use this variance type, the standard errors must be
    specified via parameter \code{S}.}
  \item{\code{"kroneker"}}{This variance type has not yet been
    implemented.}
}}

\item{...}{Additional parameters to be passed to \code{flash_backfit}.
Note that \code{nullcheck} defaults to \code{FALSE} here.}
}
\value{
A flash object, with loadings initialized from \code{LL},
  and corresponding factors initialized to zero.
}
\description{
Add a set of fixed loadings to a flash fit object
}
