% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_add_greedy}
\alias{flash_add_greedy}
\title{Fit Empirical Bayes Matrix Factorization (greedy algorithm)}
\usage{
flash_add_greedy(data, Kmax = 1, f_init = NULL, var_type = c("by_column",
  "constant"), init_fn = "udv_si", tol = 0.01, ebnm_fn = ebnm_pn,
  ebnm_param = flash_default_ebnm_param(ebnm_fn), verbose = FALSE,
  nullcheck = TRUE, seed = 123)
}
\arguments{
\item{data}{an n by p matrix or a flash data object created using \code{flash_set_data}}

\item{f_init}{if supplied, a flash object to which a single new factor is to be added}

\item{var_type}{type of variance structure to assume for residuals.}

\item{init_fn}{function to be used to initialize the factor. This function should take parameters (Y,K)
where Y is an n by p matrix of data (or a flash data object) and K is a number of factors.
It should output a list with elements (u,d,v) where u is n by K matrix
v is a p by K matrix  and d is a K vector. See \code{udv_si} for an example.
(If the input data includes missing values then this function must be able
to deal with missing values in its input matrix.)}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ebnm_fn}{function to solve the Empirical Bayes Normal Means problem}

\item{ebnm_param}{named list containing parameters to be passed to ebnm_fn when optimizing; defaults set by flash_default_ebnm_param()}

\item{verbose}{if TRUE various output progress updates will be printed}

\item{nullcheck}{flag whether to check, after running
hill-climbing updates, whether the achieved optimum is better than setting factor to 0.
If this check is performed and fails then the factor will be set to 0 in the returned fit.}

\item{seed}{a random number seed to use before running method - for reproducibility. Set to NULL if you don't want seed set.
(The seed can affect initialization when there are missing data; otherwise the algorithm is deterministic)}
}
\value{
a fitted flash object
}
\details{
This implements the greedy algorithm from Wang and Stephens.
It can be used to adds factors to an existing fit, or start from scratch.
It adds factors iteratively, at each stage adding a new factor and then optimizing it.
It is "greedy" in that it does not return to re-optimize previous factors.
The function stops when an added factor contributes nothing, or Kmax is reached.
Each new factor is intialized by applying the function `init_fn` to the residuals
after removing previously-fitted factors.
}
\examples{
l = rnorm(100)
f = rnorm(10)
Y = outer(l,f) + matrix(rnorm(1000),nrow=100)
f = flash_add_greedy(Y,10)
flash_get_ldf(f)$d #gives the weights for each factor (analogue of singular values)
# example to show how to use a different initialization function
f2 = flash_add_greedy(Y,10,init_fn = function(x,K=1){softImpute::softImpute(x,K,lambda=10)})
}
