% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_add.R
\name{flash_add_lf}
\alias{flash_add_lf}
\title{Add factors or loadings to f.}
\usage{
flash_add_lf(data, LL, FF, f_init = NULL, fixl = NULL, fixf = NULL)
}
\arguments{
\item{data}{A flash data object.}

\item{LL}{The loadings, an n by K matrix.}

\item{FF}{The factors, a p by K matrix.}

\item{fixl}{An n by K matrix of TRUE/FALSE values indicating which
elements of LL should be considered fixed and not changed during
updates. Useful for including a mean factor for example.}

\item{fixf}{A p by K matrix of TRUE/FALSE values; same as fixl but
for factors FF.}

\item{f}{A flash fit object.}
}
\value{
A flash fit object, with additional factors initialized
  using LL and FF.
}
\description{
Add factors or loadings to f.
}
\details{
The precision parameter in f is updated after adding.
}
