% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_backfit}
\alias{flash_backfit}
\title{Refines a fit of the flash model to data by "backfitting".}
\usage{
flash_backfit(data, f_init, kset = NULL, var_type = c("by_column", "by_row",
  "constant", "zero", "kroneker"), tol = 0.01, ebnm_fn = ebnm_pn,
  ebnm_param = flash_default_ebnm_param(ebnm_fn), gl = NULL,
  fixgl = FALSE, gf = NULL, fixgf = FALSE, verbose = FALSE,
  nullcheck = TRUE, maxiter = 1000)
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f_init}{The flash object to which new factors are to be added.
If \code{f_init = NULL}, then a new flash fit object is created.}

\item{kset}{The indices of factors to be optimized (NULL indicates
all factors).}

\item{var_type}{The type of variance structure to assume for
residuals.}

\item{tol}{Specifies how much the objective can change in a single
iteration to be considered not converged.}

\item{ebnm_fn}{The function used to solve the Empirical Bayes Normal
Means problem.}

\item{ebnm_param}{A named list containing parameters to be passed to
ebnm_fn when optimizing; defaults are set by
\code{flash_default_ebnm_param()}.}

\item{gl}{If nonnull, then the priors on the loadings will be fixed at
the specified values. This should be a list, with \code{gl[[k]]}
specifying the prior for the kth loading.}

\item{gf}{If nonnull, then the priors on the factors will be fixed
the specified values. This should be a list, with \code{gf[[k]]}
specifying the prior for the kth factor.}

\item{verbose}{If TRUE, various progress updates will be printed.}

\item{nullcheck}{If TRUE, then after running hill-climbing updates,
\code{flash} will check whether the achieved optimum is better than
setting the factor to 0. If the check is performed and fails then
the factor will be set to 0 in the returned fit.}

\item{maxiter}{A maximum number of iterations to perform (in the
inner loop). To perform just one iteration we suggest setting
\code{maxiter = 1} and \code{nullcheck = FALSE}.}
}
\value{
A fitted flash object.
}
\description{
Iterates through the factors of a flash object,
  updating each until convergence.
}
\examples{

l = rnorm(100) #simulate some rank 1 data
f = rnorm(10)
Y = outer(l,f) + matrix(rnorm(1000),nrow=100)
fg = flash_add_greedy(Y,10)
fb = flash_backfit(Y,fg) # refines fit from greedy by backfitting
flash_get_ldf(fb)$d

}
