% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_backfit}
\alias{flash_backfit}
\title{Refines a fit of the FLASH model to data by "backfitting"}
\usage{
flash_backfit(data, f, tol = 0.01, ash_param = list())
}
\arguments{
\item{data}{a flash data object}

\item{f}{a fitted flash object to be refined}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}
}
\value{
a fitted flash object
}
\details{
Iterates through the factors of a flash object, updating each until convergence
}
\examples{
Y = matrix(rnorm(100),nrow=5,ncol=20)
data = set_flash_data(Y)
fg = flash_greedy(data,10)
fb = flash_backfit(data,fg) # refines fit from greedy by backfitting
flash_get_sizes(fb)
}
