% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_backfit}
\alias{flash_backfit}
\title{Refines a fit of the FLASH model to data by "backfitting"}
\usage{
flash_backfit(data, f, kset = NULL, var_type = c("by_column", "constant"),
  tol = 0.01, ash_param = list(), verbose = FALSE)
}
\arguments{
\item{data}{an n by p matrix or a flash data object created using \code{set_flash_data}}

\item{f}{a fitted flash object to be refined}

\item{kset}{the indices of factors to be optimized (NULL indicates all factors)}

\item{var_type}{type of variance structure to assume for residuals.}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
a fitted flash object
}
\details{
Iterates through the factors of a flash object, updating each until convergence
}
\examples{
Y = matrix(rnorm(100),nrow=5,ncol=20)
fg = flash_greedy(Y,10)
fb = flash_backfit(Y,fg) # refines fit from greedy by backfitting
flash_get_sizes(fb)
fsi = flash_init_fn(set_flash_data(Y),"udv_si",4)
fb2 = flash_backfit(Y,fsi)
flash_get_sizes(fb2)
}
