% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flash_fill}
\alias{flash_fill}
\title{Use a flash fit to fill in missing entries}
\usage{
flash_fill(data, f)
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f}{A fitted flash object obtained from running \code{flash} on
\code{data}.}
}
\value{
A matrix with non-missing entries the same as Y, and
  missing entries imputed from the flash fit.
}
\description{
Fills in missing entries of Y by using the relevant
  entries of the estimated LDF' from the flash fit.
}
