% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_k}
\alias{flash_get_k}
\title{Get number of factors in a fit object.}
\usage{
flash_get_k(f)
}
\arguments{
\item{f}{A flash fit object.}
}
\details{
Returns the number of factors in a flash fit.
}
