% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_ldf}
\alias{flash_get_ldf}
\title{flash_get_ldf}
\usage{
flash_get_ldf(f, k = NULL)
}
\arguments{
\item{f}{a flash fit object}

\item{k}{indices of loadings/factors to be returned}
}
\value{
a list with elements
\itemize{
\item{l} a matrix whose columns contain the standardized loadings (ie norm 1)
\item{d} a vector of weights (analogous to the singular values in an svd)
\item{f} a matrix whose columns contain the standardized factors (ie norm 1)
}
These are analogous to the u, d and v returned by \code{svd}, but columns of l and f are not orthogonal.
}
\details{
returns standardized loadings, factors, and weights from a flash object.
}
