% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{flash_get_ldf}
\alias{flash_get_ldf}
\title{flash_get_ldf}
\usage{
flash_get_ldf(f, k = NULL, drop_zero_factors = TRUE)
}
\arguments{
\item{f}{A flash fit object.}

\item{k}{Indices of loadings/factors to be returned.}

\item{drop_zero_factors}{Flag whether to remove any factor/loadings
that are zero.}
}
\value{
A list with the following elements:
\itemize{
\item{l} A matrix whose columns contain the standardized loadings
  (ie norm 1).
\item{d} A vector of weights (analogous to the singular values in
  an svd).
\item{f} A matrix whose columns contain the standardized factors
  (ie norm 1).}
These are analogous to the u, d and v returned by \code{svd}, but
columns of l and f are not orthogonal.
}
\description{
flash_get_ldf
}
\details{
Returns standardized loadings, factors, and weights from a
flash object.
}
