% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{flash_get_objective}
\alias{flash_get_objective}
\title{Get value of objective function}
\usage{
flash_get_objective(data, f)
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f}{A flash fit object.}
}
\description{
Get value of objective function from data and flash fit
  object.
}
