% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_greedy}
\alias{flash_greedy}
\title{Fit the FLASH model to data by a greedy approach}
\usage{
flash_greedy(data, Kmax, nullcheck = TRUE, init_method = c("softImpute",
  "svd", "random"), tol = 0.01, ash_param = list(), verbose = FALSE)
}
\arguments{
\item{data}{an n by p matrix or a flash data object created using \code{set_flash_data}}

\item{Kmax}{the maximum number of factors to consider}

\item{nullcheck}{flag whether to check, after running
hill-climbing updates, whether the achieved optimum is better than setting factor to 0.
If this check is performed and fails then the factor will be set to 0 in the returned fit.}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
a fitted flash object
}
\details{
Fits the model by adding a factor and then optimizing it.
It is "greedy" in that it does not return to re-optimize previous factors.
Stops when an added factor contributes nothing, or Kmax is reached
}
\examples{
Y = matrix(rnorm(100),nrow=5,ncol=20)
f = flash_greedy(Y,10)
flash_get_sizes(f)
}
