% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_optimize_single_fl}
\alias{flash_optimize_single_fl}
\title{Optimize a flash factor-loading combination}
\usage{
flash_optimize_single_fl(data, f, k, tol = 0.01, ash_param = list(),
  verbose = FALSE)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash object}

\item{k}{the index of the factor/loading to optmize}

\item{tol}{a tolerance for the optimization}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
an updated flash object
}
\details{
Iteratively updates factor and loading k of f (as well as residual precision)
to convergence of objective (used in the greedy algorithm for example)
}
