% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_optimize_single_fl}
\alias{flash_optimize_single_fl}
\title{Optimize a flash factor-loading combination}
\usage{
flash_optimize_single_fl(data, f, k, var_type, nullcheck = TRUE, tol = 0.01,
  ash_param = list(), verbose = FALSE)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash object}

\item{k}{the index of the factor/loading to optimize}

\item{var_type}{type of variance structure to assume for residuals.}

\item{nullcheck}{flag whether to check, after running
hill-climbing updates, whether the achieved optimum is better than setting factor to 0.
If this check is performed and fails then the factor will be set to 0 in the returned fit.}

\item{tol}{a tolerance for the optimization}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
an updated flash object
}
\details{
Iteratively updates factor and loading k of f (as well as residual precision)
to convergence of objective (used in the greedy algorithm for example)
}
