% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_factors}
\alias{flash_plot_factors}
\title{Plot the factors from flash results.}
\usage{
flash_plot_factors(data, f, k = 1, loading_label = FALSE,
  factor_label = FALSE)
}
\arguments{
\item{data}{The flash data object.}

\item{f}{A flash fit object.}

\item{k}{Description of input argument goes here.}

\item{loading_label}{Description of input argument goes here.}

\item{factor_label}{Description of input argument goes here.}
}
\value{
List containing:

 \itemize{
  \item{\code{plot_f}} {A ggplot object for the factors.}
  \item{\code{plot_l}} {A ggplot object for the loadings}
 }
}
\details{
Plots each factor and loading as a barplot.
}
