% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_factors}
\alias{flash_plot_factors}
\title{Plot the factor/loading pairs from flash results.}
\usage{
flash_plot_factors(data, f, k = 1, loading_label = FALSE,
  factor_label = FALSE)
}
\arguments{
\item{data}{The flash data object.}

\item{f}{The flash fit object.}

\item{k}{The indices of the factor/loading pairs to be plotted.}

\item{loading_label}{The label to use for the loading plots.}

\item{factor_label}{The label to use for the factor plots.}
}
\value{
List containing:
  \describe{
    \item{\code{plot_f}}{A ggplot object for the factors.}
    \item{\code{plot_l}}{A ggplot object for the loadings.}
  }
}
\description{
Plot the factor/loading pairs from flash results.
}
\details{
Plots each factor and loading as a barplot.
}
