% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_pve}
\alias{flash_plot_pve}
\title{Scree plot.}
\usage{
flash_plot_pve(f, main = "Scree plot of PVE for each factor")
}
\arguments{
\item{f}{The flash fit object.}

\item{main}{Description of input argument goes here.}
}
\value{
A ggplot plot object.
}
\description{
Create a scree plot giving proportion of variance
explained by each factor.
}
