% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{flash_plot_pve}
\alias{flash_plot_pve}
\title{Scree plot}
\usage{
flash_plot_pve(f, main = "Scree plot of PVE for each factor",
  drop_zero_factors = TRUE)
}
\arguments{
\item{f}{A flash fit object.}

\item{main}{The main caption to use for the plot.}

\item{drop_zero_factors}{If \code{TRUE}, then any factor/loadings
that are zero will be removed.}
}
\value{
A \pkg{ggplot} plot object.
}
\description{
Create a scree plot giving the proportion of variance
explained by each factor.
}
