% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash_r1}
\alias{flash_r1}
\title{Fit the rank1 FLASH model to data}
\usage{
flash_r1(data, init_method = c("svd", "random"), tol = 0.01,
  ash_param = list())
}
\arguments{
\item{data}{a flash data object}

\item{init_method}{specifies how to initialize the factors. Options include svd on data matrix, or random (N(0,1))}

\item{tol}{specify how much objective can change in a single iteration to be considered not converged}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}
}
\value{
a fitted flash object
}
\examples{
Y = matrix(rnorm(100),nrow=5,ncol=20)
data = set_flash_data(Y)
f = flash_r1(data)
flash_get_sizes(f)
}
