% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_update_precision}
\alias{flash_update_precision}
\title{Update precision parameter}
\usage{
flash_update_precision(Y, f)
}
\arguments{
\item{Y}{an n times p data matrix}

\item{f}{a flash object}
}
\value{
an updated flash object
}
\details{
Updates precision estimate to increase the objective F
}
