% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_precision.R
\name{flash_update_precision}
\alias{flash_update_precision}
\title{Update precision parameter.}
\usage{
flash_update_precision(data, f, var_type = c("by_column", "by_row",
  "constant", "zero", "kroneker"))
}
\arguments{
\item{data}{A flash data object.}

\item{f}{A flash object.}

\item{var_type}{Indicates what type of variance structure to assume.}
}
\value{
An updated flash object.
}
\description{
Updates precision estimate to increase the objective F.
}
