% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_precision.R
\name{flash_update_precision}
\alias{flash_update_precision}
\title{Update precision parameter}
\usage{
flash_update_precision(data, f, var_type = c("by_column", "by_row",
  "constant", "zero", "kroneker"))
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f}{A flash object.}

\item{var_type}{The type of variance structure to assume for
residuals. Options include:
\describe{
  \item{\code{"by_column"}}{Residuals in any given column are
    assumed to have the same variance.}
  \item{\code{"by_row"}}{Residuals in any given row have the
    same variance.}
  \item{\code{"constant"}}{All residuals are assumed to have the
    same variance.}
  \item{\code{"zero"}}{The variance of the residuals is fixed. To
    use this variance type, the standard errors must be
    specified via parameter \code{S} when using
    \code{flash_set_data} to set the flash data object.}
  \item{\code{"kroneker"}}{This variance type has not yet been
    implemented.}
}}
}
\value{
An updated flash object.
}
\description{
Updates the estimated precision to increase the value of
  the objective function.
}
