% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_update_single_factor}
\alias{flash_update_single_factor}
\title{Update a flash factor}
\usage{
flash_update_single_factor(data, f, k, ebnm_fn = ebnm_ash,
  ebnm_param = flash_default_ash_param())
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{k}{the index of the loading to update}

\item{ebnm_fn}{function to solve the Empirical Bayes normal means problem}

\item{ebnm_param}{parameters to be passed to ebnm_fn when optimizing}
}
\value{
an updated flash object
}
\details{
Updates factor k of f to increase the objective F.
Updates only the factor, once (not the loading)
}
