% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_update_single_factor}
\alias{flash_update_single_factor}
\title{Update a flash factor}
\usage{
flash_update_single_factor(data, f, k)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash object}

\item{k}{the index of the factor to update}
}
\value{
an updated flash object
}
\details{
Updates factor k of f to increase the objective F updates only the factor, once (not the loading)
}
