% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_update_single_loading}
\alias{flash_update_single_loading}
\title{Update a flash loading}
\usage{
flash_update_single_loading(data, f, k, ash_param = list())
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{k}{the index of the loading to update}

\item{ash_param}{parameters to be passed to ashr when optimizing; defaults set by flash_default_ash_param()}
}
\value{
an updated flash object
}
\details{
Updates loading k of f to increase the objective F
Updates only the loading, once (not the factor)
}
