% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flash_update_single_loading}
\alias{flash_update_single_loading}
\title{Update a flash loading}
\usage{
flash_update_single_loading(data, f, k, ebnm_fn = ebnm_ash,
  ebnm_param = list())
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{k}{the index of the loading to update}

\item{ebnm_fn}{function to solve the Empirical Bayes normal means problem}

\item{ebnm_param}{parameters to be passed to ebnm_fn when optimizing; defaults set by flash_default_ebnm_param()}
}
\value{
an updated flash object
}
\details{
Updates loading k of f to increase the objective F
Updates only the loading, once (not the factor)
}
