% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_zero_out_factor}
\alias{flash_zero_out_factor}
\title{zero out a factor from f}
\usage{
flash_zero_out_factor(data, f, k = 1)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}
}
\details{
the specified factor is made to be 0, effectively reducing the rank by 1
}
