% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{flash_zero_out_factor}
\alias{flash_zero_out_factor}
\title{zero out a factor from f}
\usage{
flash_zero_out_factor(data, f, k = 1)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash fit object}

\item{k}{index of factor/loading to zero out}
}
\details{
The factor and loadings of the kth factor of f are made to be zero.
(Except for elements of the factor/loading that are designated to be fixed)
This effectively reduces the rank by 1, although the zero factor/loading is kept in f
so the number and indexing of factor/loading matrices in f remains the same
}
