% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_fit_utils.R
\name{flash_zero_out_factor}
\alias{flash_zero_out_factor}
\title{Zero out factor from flash object}
\usage{
flash_zero_out_factor(data, f_init, k)
}
\arguments{
\item{data}{An n by p matrix or a flash data object created using
\code{flash_set_data}.}

\item{f_init}{The flash object or flash fit object to which new
factors are to be added. If \code{f_init = NULL}, then a new flash
object is created.}

\item{k}{The index of the factor/loading pair to zero out.}
}
\value{
The flash object with the kth factor/loading zeroed out.
}
\description{
The factor and loadings of the kth factor of \code{f}
  are made to be zero (except for elements of the factor/loading that
  are designated to be fixed). This effectively reduces the rank by 1,
  but the zero factor/loading is retained in \code{f} so that
  the number and indexing of factor/loading matrices in \code{f}
  remain the same.
}
