% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_utils.R
\name{flash_zero_out_factor}
\alias{flash_zero_out_factor}
\title{Zero out factor from flash object}
\usage{
flash_zero_out_factor(f, k)
}
\arguments{
\item{f}{A flash fit object.}

\item{k}{The index of the factor/loading pair to zero out.}
}
\description{
The factor and loadings of the kth factor of \code{f}
  are made to be zero (except for elements of the factor/loading that
  are designated to be fixed). This effectively reduces the rank by 1,
  but the zero factor/loading is retained in \code{f} so that
  the number and indexing of factor/loading matrices in \code{f}
  remains the same.
}
