% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashcard.R
\name{flashcard}
\alias{flashcard}
\title{Create flashcards}
\usage{
flashcard(x, termsfirst = TRUE, package = TRUE, theme = "moon", file = NULL)
}
\arguments{
\item{x}{Name of pre-existing flashcard deck or path and name of CSV file
containing terms and descriptions}

\item{termsfirst}{Logical indicating whether to show terms first (TRUE) or
descriptions first (FALSE)}

\item{package}{Logical indicating whether to include package name in term}

\item{theme}{Name of reveal.js theme to use for flashcards}

\item{file}{Path and file name used to save flashcard deck locally (must
save as HTML)}
}
\value{
An HTML file of terms and descriptions rendered in the RStudio viewer or
web browser.
}
\description{
The \code{flashcard()} function generates a set of flashcards with randomly
ordered pairs of terms and descriptions. The function outputs reveal.js
presentation as an HTML file. If running in RStudio, the flashcards are
output to the viewer. Otherwise, they are output to a web browser.
}
\examples{
\dontrun{
# Display terms then descriptions
flashcard(data_types)

# Display descriptions then terms
flashcard(data_types, termsfirst = FALSE)

# Display custom CSV file of terms and descriptions.
# If package information is not included, set `package = FALSE`.
flashcard(file = "data/operators.csv", package = FALSE)
}
}
