% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{perform_nullcheck}
\alias{perform_nullcheck}
\title{Zeros out factors when that improves the objective}
\usage{
perform_nullcheck(data, f, kset, var_type, verbose)
}
\arguments{
\item{data}{a flash data object}

\item{f}{a flash object}

\item{kset}{the indices of the factor/loading to check}

\item{var_type}{type of variance structure to assume for residuals.}

\item{verbose}{if TRUE various output progress updates will be printed}
}
\value{
a flash object
}
\description{
Zeros out factors when that improves the objective
}
\details{
Sometimes zeroing out a factor can improve the objective.
This function iterates over factors with indices in kset
and checks whether zeroing it out will improve the objective; if so then that factor
is set to 0 (and precision is updated). Returns the final flash fit object obtained when this iterative process stops
(ie a complete pass is performed with no factor being zerod)
}
