% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_one_factor}
\alias{plot_one_factor}
\title{Factor plot.}
\usage{
plot_one_factor(f, pve, k, f_labels = NA, y_lab = "factor values")
}
\arguments{
\item{f}{Factor to plot.}

\item{pve}{PVE for this factor.}

\item{k}{The order of the factor.}

\item{f_labels}{The labels for the factor.}

\item{y_lab}{The name of the Y axis.}

\item{P}{The length of this factor.}
}
\value{
list of factor, loading and variance of noise matrix
 \itemize{
  \item{\code{plot_f}} {is a ggplot object for the factors}
 }
}
\description{
Factor plot.
}
\details{
Plots the factors in a barplot.
}
