% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.flash}
\alias{print.flash}
\title{Print summary of flash object}
\usage{
\method{print}{flash}(x, ...)
}
\arguments{
\item{x}{A flash object.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
\code{print} method for class \code{'flash'}.
}
