% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_data.R
\name{set_flash_data}
\alias{set_flash_data}
\title{set up data for reading into flash}
\usage{
set_flash_data(Y)
}
\arguments{
\item{Y}{an n by p data matrix}
}
\value{
a flash data object
}
\details{
Y can have missing data, but no column or row can be entirely missing
The flash data object contains flags for dealing with missing data
and a (naively) imputed version of the original data matrix so that
i) some of the initializatinon methods used by flash (eg svd) do not fail
ii) data$Y * data$missing is 0 if the original data were missing
}
