% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{tbackfitting}
\alias{tbackfitting}
\title{Perform backfitting starting at output of \code{tgreedy}.}
\usage{
tbackfitting(Y, factor_list, sigma_est, maxiter_bf = 100, tol_bf = 10^-6,
  maxiter_vem = 100, var_type = c("homoscedastic", "kronecker"),
  mixcompdist = "normal", alpha = 0, beta = 0, nullweight = 10,
  known_factors = NULL, known_modes = NULL, tol_r1 = 10^-3)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{factor_list}{A list of matrices with the same number of
columns. These are the starting values for the backfitting
algorithm. The intended starting values are can be obtained
from \code{\link{tgreedy}}.}

\item{sigma_est}{Either a vector of estimated precisions (when
\code{var_type = "homoscedastic"}) or a list of matrices whose
columns are estimated precisions (when \code{var_type =
"kronecker"}).}

\item{maxiter_bf}{A positive integer. The maximum number of
backfitting steps to perform.}

\item{tol_bf}{A positive numeric. The stopping criterion for the
backfitting algorithm.}

\item{maxiter_vem}{A positive integer. The maximum number of steps
in each VEM algorithm to perform at each iteration of the
backfitting algorithm.}

\item{var_type}{A string. What variance model should we assume?
Options are homoscedastic noise (\code{"homoscedastic"}) or
Kronecker structured variance (\code{kronecker}).}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of zero.}

\item{known_factors}{A list of known factors for the modes
indicated in \code{known_modes}. Defaults to \code{NULL}, where
all factors are assumed to be unknown.}

\item{known_modes}{A vector of integers. The modes that are
known. Should be the same length as \code{known_factors}.}

\item{tol_r1}{A positive numeric. The tolerance for the rank 1 runs
of tflash.}
}
\description{
Perform backfitting starting at output of \code{tgreedy}.
}
\author{
David Gerard
}

