% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflash.R
\name{tflash}
\alias{tflash}
\title{Tensor Factor Loading Adaptive SHrinkage (T-FLASH).}
\usage{
tflash(Y, tol = 10^-5, itermax = 100, alpha = 0, beta = 0,
  mixcompdist = "normal")
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}
}
\description{
Fits a rank-1 tensor mean model with a homoscedastic error term. It
does this via a variational Bayesian approach with a unimodal prior
on the components of the mean tensor.
}
\author{
David Gerard
}

