% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tflash_kron}
\alias{tflash_kron}
\title{Tensor FLASH assuming a diagonal Kronecker structured covariance model.}
\usage{
tflash_kron(Y, tol = 10^-5, itermax = 100, alpha = 0, beta = 0,
  mixcompdist = "normal")
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}
}
\description{
Tensor FLASH assuming a diagonal Kronecker structured covariance model.
}
\author{
David Gerard
}

