% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbackfit.R
\name{tgreedy}
\alias{tgreedy}
\title{Iteratively do T-FLASH on the residuals.}
\usage{
tgreedy(Y, k = max(dim(Y)), tol = 10^-5, itermax = 100, alpha = 0,
  beta = 0, mixcompdist = "normal", var_type = c("homoscedastic",
  "kronecker"), nullweight = 1, print_update = FALSE)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{k}{The maximum cp-rank of the mean tensor.}

\item{tol}{A positive numeric. The stopping criterion for the VEM.}

\item{itermax}{A positive integer. The maximum number of iterations
to run the VEM}

\item{alpha}{A non-negative numeric. The prior shape parameter for
the variance. Defaults to zero.}

\item{beta}{A non-negative numeric. The prior rate parameter for
the variance. Defaults to zero.}

\item{mixcompdist}{The mixing distribution to assume. Defaults to
normal. Options are those available in the \code{ashr} package.}

\item{var_type}{A string. What variance model should we assume?
Options are homoscedastic noise (\code{"homoscedastic"}) or
Kronecker structured variance (\code{kronecker}).}

\item{nullweight}{A numeric greater than or equal to 1. The penalty
term on the probability of zero.}

\item{print_update}{A logical. Should we print notifications on how
far along the optimization is?}
}
\value{
\code{factor_list}: A list of matrices of
    numerics. \code{factor_list[[i]][, j]} contains the \eqn{j}th
    factor of the \eqn{i}th mode.

    \code{sigma_est}: If \code{var_type = "homoscedastic"}, then
    \code{sigma_est} is a vector of numerics. \code{sigma_est[i]}
    is the estimate of the precision during the \eqn{i} iteration
    of the greedy algorithm. Only the last one (if that) should
    actually be used for any sort of precision estimate.

    If \code{var_type = "kronecker"}, then \code{sigma_est} is a
    list of matices. \code{sigma_est[[i]][, j]} is the estimate of
    the variances for the \eqn{j}th mode during the \eqn{i}th run
    of the greedy algorithm. Only the final columns (if those) of
    these matrices should actually be used as any sort of precision
    estimate.

    \code{rank_final} A non-negative integer. The final estimated
    cp-rank of the mean.
}
\description{
This function will estimate a rank-1 tensor using independent
unimodal priors on the components via a variational EM
algorithm. It will subtract off the posterior mean from the data
array, then repeat the estimation procedure on the residuals. It
continues to do this until the maximum cp-rank \code{k} is reached
or when all of the components are estimated to be 0.
}
\author{
David Gerard
}
\seealso{
\code{\link{tflash}} for fitting the rank-1 mean tensor
    model.
}

