% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tinit_kron_components}
\alias{tinit_kron_components}
\title{Obtain initial estimates of each mode's components.}
\usage{
tinit_kron_components(Y, which_na = NULL, start = c("first_sv", "random"),
  known_factors = NULL, known_modes = NULL)
}
\arguments{
\item{Y}{An array of numerics. The data.}

\item{which_na}{Either NULL or an array the same dimension as
\code{Y} indicating if an element of \code{Y} is missing
(\code{TRUE}) or observed (\code{FALSE}).}

\item{start}{How should we choose the starting values? Either using
the first singular vector along each mode (\code{"first_sv"})
or randomly (\code{"random"}).}

\item{known_factors}{A list of known factors for the modes
indicated in \code{known_modes}. Defaults to \code{NULL}, where
all factors are assumed to be unknown.}

\item{known_modes}{A vector of integers. The modes that are
known. Should be the same length as \code{known_factors}.}
}
\value{
\code{ex_list} A list of vectors of the starting expected
    values of each component.


    \code{ex2_list} A vector of starting values of \eqn{E[x^2]}
}
\description{
The inital values of each component are taken to be proportional to
the first singular vector of each mode-k matricization of the data
array.
}
\author{
David Gerard
}

