% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkron.R
\name{tinit_kron_components}
\alias{tinit_kron_components}
\title{Obtain initial estimates of each mode's components.}
\usage{
tinit_kron_components(Y, which_na = NULL, start = c("first_sv", "random"))
}
\arguments{
\item{Y}{An array of numerics.}

\item{which_na}{Either NULL or an array the same dimension as
\code{Y} indicating if an element of \code{Y} is missing
(\code{TRUE}) or observed (\code{FALSE}).}
}
\value{
\code{ex_list} A list of vectors of the starting expected
    values of each component.


    \code{ex2_list} A vector of starting values of \eqn{E[x^2]}
}
\description{
The inital values of each component are taken to be proportional to
the first singular vector of each mode-k matricization of the data
array.
}
\author{
David Gerard
}

