% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_random}
\alias{udv_random}
\title{udv_random}
\usage{
udv_random(Y, K = 1)
}
\arguments{
\item{Y}{an n by p matrix}

\item{K}{number of factors}
}
\value{
a list with components (u,d,v), with elements of u and v iid N(0,1)
}
\details{
provides a random initialization of factors
}
