% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_si}
\alias{udv_si}
\title{udv_si}
\usage{
udv_si(Y, K = 1)
}
\arguments{
\item{Y}{an n by p matrix}

\item{K}{number of factors to use}
}
\value{
a list with components (u,d,v)
}
\details{
provides a simple wrapper to \code{softImpute} to provide a rank 1 initialization
}
