% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{udv_svd}
\alias{udv_svd}
\title{udv_svd}
\usage{
udv_svd(Y, K = 1)
}
\arguments{
\item{Y}{an n by p matrix}

\item{K}{number of factors to use}
}
\value{
a list with components (u,d,v)
}
\description{
udv_svd
}
\details{
provides a simple wrapper to svd
}
